Components (submodels)

Spacetimes

SymBoltz.metric — Function

metric(; name = :g, kwargs...)

Create a symbolic component for the perturbed FLRW spacetime metric in the conformal Newtonian gauge with sign signature diag(-1, +1, +1, +1).

g = SymBoltz.metric()

Equations

equations(g)

\[ \begin{align*} z\left( \tau \right) &= -1 + \frac{1}{a\left( \tau \right)} \\ \dot{z}\left( \tau \right) &= \frac{\mathrm{d} z\left( \tau \right)}{\mathrm{d}\tau} \\ \mathscr{E}\left( \tau \right) &= \frac{\frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau}}{a\left( \tau \right)} \\ E\left( \tau \right) &= \frac{\mathscr{E}\left( \tau \right)}{a\left( \tau \right)} \\ \mathscr{H}\left( \tau \right) &= 3.2408 \cdot 10^{-18} h \mathscr{E}\left( \tau \right) \\ H\left( \tau \right) &= 3.2408 \cdot 10^{-18} h E\left( \tau \right) \end{align*} \]

Gravitational theories

SymBoltz.general_relativity — Function

general_relativity(g; name = :G, kwargs...)

Create a symbolic component for the general relativistic (GR) theory of gravity in the spacetime with the metric g.

G = SymBoltz.general_relativity(g)

Equations

equations(G)

\[ \begin{align*} \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} &= \left( a\left( \tau \right) \right)^{2} \sqrt{\frac{8}{3} \rho\left( \tau \right) \pi} \\ \mathtt{F_1}\left( \tau \right) &= 0 \\ \mathtt{F_2}\left( \tau \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} + \frac{ - \left( \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} \right)^{2} \left( 1 + \frac{ - 3 P\left( \tau \right)}{\rho\left( \tau \right)} \right)}{2 a\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) &= \frac{ - \frac{4}{3} \left( a\left( \tau \right) \right)^{2} \mathtt{\delta\rho}\left( \tau, k \right) \pi}{\mathscr{E}\left( \tau \right)} + \frac{ - k^{2} \Phi\left( \tau, k \right)}{3 \mathscr{E}\left( \tau \right)} - \Psi\left( \tau, k \right) \mathscr{E}\left( \tau \right) \\ k^{2} \left( \Phi\left( \tau, k \right) - \Psi\left( \tau, k \right) \right) &= 12 \left( a\left( \tau \right) \right)^{2} \Pi\left( \tau, k \right) \pi \\ \mathtt{\dot{\Psi}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} \Psi\left( \tau, k \right) \\ \mathtt{\dot{\Phi}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \end{align*} \]

Initialization equations

initialization_equations(G)

\[ \begin{align*} \end{align*} \]

SymBoltz.brans_dicke — Function

brans_dicke(g; name = :G, kwargs...)

Create a symbolic component for the Brans-Dicke (BD) theory of gravity in the spacetime with the metric g.

G = SymBoltz.brans_dicke(g)

Equations

equations(G)

\[ \begin{align*} \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} &= \frac{8 \left( a\left( \tau \right) \right)^{2} \left( - 3 P\left( \tau \right) + \rho\left( \tau \right) \right) \pi}{3 + 2 \omega} + \frac{ - 2 \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{a\left( \tau \right)} \\ G\left( \tau \right) &= \frac{4 + 2 \omega}{\phi\left( \tau \right) \left( 3 + 2 \omega \right)} \\ \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} &= \frac{ - \frac{1}{2} a\left( \tau \right) \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} + \sqrt{\frac{\frac{8}{3} \left( a\left( \tau \right) \right)^{4} \rho\left( \tau \right) \pi}{\phi\left( \tau \right)} + \left( \frac{\frac{1}{2} a\left( \tau \right) \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} \right)^{2} + \frac{1}{6} \left( \frac{a\left( \tau \right) \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} \right)^{2} \omega} \\ \mathtt{F_1}\left( \tau \right) &= 0 \\ \mathtt{F_2}\left( \tau \right) &= \frac{ - \left( \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} \right)^{2}}{2 a\left( \tau \right)} + \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} + \frac{4 \left( a\left( \tau \right) \right)^{3} P\left( \tau \right) \pi}{\phi\left( \tau \right)} + \frac{\frac{1}{2} a\left( \tau \right) \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} + \frac{\frac{1}{2} \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} + \frac{1}{4} \left( \frac{\frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} \right)^{2} a\left( \tau \right) \omega \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) &= \frac{3 \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{\delta\phi}\left( \tau, k \right) \mathscr{E}\left( \tau \right) + \left( k^{2} + 3 \left( \mathscr{E}\left( \tau \right) \right)^{2} \right) \mathtt{\delta\phi}\left( \tau, k \right) - 2 k^{2} \Phi\left( \tau, k \right) \phi\left( \tau \right) + \frac{1}{2} \left( - 2 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{\delta\phi}\left( \tau, k \right) + \left( \frac{\frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} \right)^{2} \mathtt{\delta\phi}\left( \tau, k \right) \right) \omega - 8 \left( a\left( \tau \right) \right)^{2} \left( \mathtt{\delta\rho}\left( \tau, k \right) + 2 \Psi\left( \tau, k \right) \rho\left( \tau \right) \right) \pi}{3 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} + 6 \mathscr{E}\left( \tau \right) \phi\left( \tau \right)} \\ \Psi\left( \tau, k \right) &= \frac{ - \mathtt{\delta\phi}\left( \tau, k \right)}{\phi\left( \tau \right)} + \Phi\left( \tau, k \right) + \frac{ - 12 \left( a\left( \tau \right) \right)^{2} \Pi\left( \tau, k \right) \pi}{k^{2} \phi\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{\delta\phi}\left( \tau, k \right) &= \frac{ - 8 \left( a\left( \tau \right) \right)^{2} \left( - \mathtt{\delta\rho}\left( \tau, k \right) + 3 \mathtt{{\delta}P}\left( \tau, k \right) \right) \pi}{3 + 2 \omega} + 2 \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \Psi\left( \tau, k \right) + \left( \frac{\mathrm{d}}{\mathrm{d}\tau} \Psi\left( \tau, k \right) + 3 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \right) \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} - 2 \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{\delta\phi}\left( \tau, k \right) \mathscr{E}\left( \tau \right) - k^{2} \mathtt{\delta\phi}\left( \tau, k \right) + 4 \Psi\left( \tau, k \right) \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \mathscr{E}\left( \tau \right) \\ \mathtt{\dot{\Psi}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} \Psi\left( \tau, k \right) \\ \mathtt{\dot{\Phi}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \end{align*} \]

Initialization equations

initialization_equations(G)

\[ \begin{align*} \mathtt{\delta\phi}\left( \tau, k \right) &= 0 \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{\delta\phi}\left( \tau, k \right) &= 0 \end{align*} \]

Species

Generic

SymBoltz.species_constant_eos — Function

species_constant_eos(g, w, cₛ² = w, ẇ = 0, _σ = 0; θinteract = false, kwargs...)

Create a symbolic component for a particle species with equation of state w ~ P/ρ in the spacetime with the metric g.

using SymBoltz: τ
@variables w(τ)
s = SymBoltz.species_constant_eos(g, w)

Equations

equations(s)

\[ \begin{align*} w\left( \tau \right) &= w\left( \tau \right) \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \delta\left( \tau, k \right) &= \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau, k \right) - 3 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau, k \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \theta\left( \tau, k \right) &= \mathtt{{\theta}interaction}\left( \tau, k \right) + \frac{k^{2} \delta\left( \tau, k \right) \mathtt{c_s^2}\left( \tau, k \right)}{1 + w\left( \tau \right)} + k^{2} \Psi\left( \tau, k \right) - k^{2} \sigma\left( \tau, k \right) - \left( 1 - 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau, k \right) \\ u\left( \tau, k \right) &= \frac{\theta\left( \tau, k \right)}{k} \\ \mathtt{\dot{u}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} u\left( \tau, k \right) \\ \sigma\left( \tau, k \right) &= 0 \\ \mathtt{{\theta}interaction}\left( \tau, k \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(s)

\[ \begin{align*} \delta\left( \tau, k \right) &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \end{align*} \]

Relativistic

SymBoltz.radiation — Function

radiation(g; name = :r, kwargs...)

Create a particle species for radiation (with equation of state w ~ 1/3) in the spacetime with metric g.

r = SymBoltz.radiation(g)

Equations

equations(r)

\[ \begin{align*} T\left( \tau \right) &= \frac{\mathtt{T{_0}}}{a\left( \tau \right)} \\ w\left( \tau \right) &= \frac{1}{3} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= 0.11937 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( \tau \right) &= 8.3776 \rho\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \delta\left( \tau, k \right) &= \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau, k \right) - 3 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau, k \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \theta\left( \tau, k \right) &= \mathtt{{\theta}interaction}\left( \tau, k \right) + \frac{k^{2} \delta\left( \tau, k \right) \mathtt{c_s^2}\left( \tau, k \right)}{1 + w\left( \tau \right)} + k^{2} \Psi\left( \tau, k \right) - k^{2} \sigma\left( \tau, k \right) + \left( -1 + 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau, k \right) \\ u\left( \tau, k \right) &= \frac{\theta\left( \tau, k \right)}{k} \\ \mathtt{\dot{u}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} u\left( \tau, k \right) \\ \sigma\left( \tau, k \right) &= 0 \\ \mathtt{{\theta}interaction}\left( \tau, k \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(r)

\[ \begin{align*} \delta\left( \tau, k \right) &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.photons — Function

photons(g; polarization = true, lmax = 6, name = :γ, kwargs...)

Create a particle species for photons in the spacetime with metric g.

γ = SymBoltz.photons(g; lmax = 6)

Equations

equations(γ)

\[ \begin{align*} \mathtt{\Omega{_0}} &= \frac{1.4121 \cdot 10^{-41} \mathtt{T{_0}}^{4}}{3.1508 \cdot 10^{-35} h^{2}} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{F0}\left( \tau, k \right) &= 4 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) - F\_{1}\left( \tau, k \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{1}\left( \tau, k \right) &= \frac{ - \frac{4}{3} \mathtt{\dot{\kappa}}\left( \tau \right) \left( \mathtt{{\theta}b}\left( \tau, k \right) - \theta\left( \tau, k \right) \right)}{k} + \frac{1}{3} \left( - 2 F\_{2}\left( \tau, k \right) + 4 \Psi\left( \tau, k \right) + \mathtt{F0}\left( \tau, k \right) \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{2}\left( \tau, k \right) &= \frac{1}{5} \left( 2 F\_{1}\left( \tau, k \right) - 3 F\_{3}\left( \tau, k \right) \right) k + \left( F\_{2}\left( \tau, k \right) - \frac{1}{10} \Pi\left( \tau, k \right) \right) \mathtt{\dot{\kappa}}\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{3}\left( \tau, k \right) &= \frac{1}{7} \left( 3 F\_{2}\left( \tau, k \right) - 4 F\_{4}\left( \tau, k \right) \right) k + F\_{3}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{4}\left( \tau, k \right) &= \frac{1}{9} \left( 4 F\_{3}\left( \tau, k \right) - 5 F\_{5}\left( \tau, k \right) \right) k + F\_{4}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{5}\left( \tau, k \right) &= \frac{1}{11} \left( 5 F\_{4}\left( \tau, k \right) - 6 F\_{6}\left( \tau, k \right) \right) k + F\_{5}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{6}\left( \tau, k \right) &= F\_{5}\left( \tau, k \right) k + F\_{6}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) - 7 F\_{6}\left( \tau, k \right) \mathscr{E}\left( \tau \right) \\ \delta\left( \tau, k \right) &= \mathtt{F0}\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{3}{4} F\_{1}\left( \tau, k \right) k \\ \sigma\left( \tau, k \right) &= \frac{1}{2} F\_{2}\left( \tau, k \right) \\ \Pi\left( \tau, k \right) &= F\_{2}\left( \tau, k \right) + G\_{2}\left( \tau, k \right) + \mathtt{G0}\left( \tau, k \right) \\ \mathtt{\dot{\Pi}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} \Pi\left( \tau, k \right) \\ \mathtt{c_s^2}\left( \tau, k \right) &= \frac{1}{3} \\ \mathtt{{\Theta}0}\left( \tau, k \right) &= \frac{1}{4} \mathtt{F0}\left( \tau, k \right) \\ \Theta\_{1}\left( \tau, k \right) &= \frac{1}{4} F\_{1}\left( \tau, k \right) \\ \Theta\_{2}\left( \tau, k \right) &= \frac{1}{4} F\_{2}\left( \tau, k \right) \\ \Theta\_{3}\left( \tau, k \right) &= \frac{1}{4} F\_{3}\left( \tau, k \right) \\ \Theta\_{4}\left( \tau, k \right) &= \frac{1}{4} F\_{4}\left( \tau, k \right) \\ \Theta\_{5}\left( \tau, k \right) &= \frac{1}{4} F\_{5}\left( \tau, k \right) \\ \Theta\_{6}\left( \tau, k \right) &= \frac{1}{4} F\_{6}\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{G0}\left( \tau, k \right) &= - G\_{1}\left( \tau, k \right) k + \mathtt{\dot{\kappa}}\left( \tau \right) \left( \mathtt{G0}\left( \tau, k \right) - \frac{1}{2} \Pi\left( \tau, k \right) \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} G\_{1}\left( \tau, k \right) &= G\_{1}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) + \frac{1}{3} \left( - 2 G\_{2}\left( \tau, k \right) + \mathtt{G0}\left( \tau, k \right) \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} G\_{2}\left( \tau, k \right) &= \frac{1}{5} \left( 2 G\_{1}\left( \tau, k \right) - 3 G\_{3}\left( \tau, k \right) \right) k + \left( G\_{2}\left( \tau, k \right) - \frac{1}{10} \Pi\left( \tau, k \right) \right) \mathtt{\dot{\kappa}}\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} G\_{3}\left( \tau, k \right) &= \frac{1}{7} \left( 3 G\_{2}\left( \tau, k \right) - 4 G\_{4}\left( \tau, k \right) \right) k + G\_{3}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} G\_{4}\left( \tau, k \right) &= \frac{1}{9} \left( 4 G\_{3}\left( \tau, k \right) - 5 G\_{5}\left( \tau, k \right) \right) k + G\_{4}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} G\_{5}\left( \tau, k \right) &= \frac{1}{11} \left( 5 G\_{4}\left( \tau, k \right) - 6 G\_{6}\left( \tau, k \right) \right) k + G\_{5}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} G\_{6}\left( \tau, k \right) &= G\_{5}\left( \tau, k \right) k + G\_{6}\left( \tau, k \right) \mathtt{\dot{\kappa}}\left( \tau \right) - 7 G\_{6}\left( \tau, k \right) \mathscr{E}\left( \tau \right) \\ T\left( \tau \right) &= \frac{\mathtt{T{_0}}}{a\left( \tau \right)} \\ w\left( \tau \right) &= \frac{1}{3} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= 0.11937 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( \tau \right) &= 8.3776 \rho\left( \tau \right) \end{align*} \]

Initialization equations

initialization_equations(γ)

\[ \begin{align*} \mathtt{F0}\left( \tau, k \right) &= - 2 \Psi\left( \tau, k \right) \\ F\_{1}\left( \tau, k \right) &= \frac{\frac{2}{3} k \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \\ F\_{2}\left( \tau, k \right) &= \frac{ - \frac{8}{15} F\_{1}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ F\_{3}\left( \tau, k \right) &= \frac{ - \frac{3}{7} F\_{2}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ F\_{4}\left( \tau, k \right) &= \frac{ - \frac{4}{9} F\_{3}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ F\_{5}\left( \tau, k \right) &= \frac{ - \frac{5}{11} F\_{4}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ F\_{6}\left( \tau, k \right) &= \frac{ - \frac{6}{13} F\_{5}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ \mathtt{G0}\left( \tau, k \right) &= \frac{5}{16} F\_{2}\left( \tau, k \right) \\ G\_{1}\left( \tau, k \right) &= \frac{ - \frac{1}{16} F\_{2}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{2}\left( \tau, k \right) &= \frac{1}{16} F\_{2}\left( \tau, k \right) \\ G\_{3}\left( \tau, k \right) &= \frac{ - \frac{3}{7} G\_{2}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{4}\left( \tau, k \right) &= \frac{ - \frac{4}{9} G\_{3}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{5}\left( \tau, k \right) &= \frac{ - \frac{5}{11} G\_{4}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{6}\left( \tau, k \right) &= \frac{ - \frac{6}{13} G\_{5}\left( \tau, k \right) k}{\mathtt{\dot{\kappa}}\left( \tau \right)} \end{align*} \]

Non-relativistic

SymBoltz.matter — Function

matter(g; name = :m, kwargs...)

Create a particle species for matter (with equation of state w ~ 0) in the spacetime with metric g.

m = SymBoltz.matter(g)

Equations

equations(m)

\[ \begin{align*} w\left( \tau \right) &= 0 \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \delta\left( \tau, k \right) &= \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau, k \right) - 3 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau, k \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \theta\left( \tau, k \right) &= \mathtt{{\theta}interaction}\left( \tau, k \right) + \frac{k^{2} \delta\left( \tau, k \right) \mathtt{c_s^2}\left( \tau, k \right)}{1 + w\left( \tau \right)} + k^{2} \Psi\left( \tau, k \right) - k^{2} \sigma\left( \tau, k \right) - \left( 1 - 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau, k \right) \\ u\left( \tau, k \right) &= \frac{\theta\left( \tau, k \right)}{k} \\ \mathtt{\dot{u}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} u\left( \tau, k \right) \\ \sigma\left( \tau, k \right) &= 0 \\ \mathtt{{\theta}interaction}\left( \tau, k \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(m)

\[ \begin{align*} \delta\left( \tau, k \right) &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.cold_dark_matter — Function

cold_dark_matter(g; name = :c, kwargs...)

Create a particle species for cold dark matter in the spacetime with metric g.

c = SymBoltz.cold_dark_matter(g)

Equations

equations(c)

\[ \begin{align*} \mathtt{c_s^2}\left( \tau, k \right) &= 0 \\ w\left( \tau \right) &= 0 \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= 0.11937 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( \tau \right) &= 8.3776 \rho\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \delta\left( \tau, k \right) &= \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau, k \right) - 3 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau, k \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \theta\left( \tau, k \right) &= \mathtt{{\theta}interaction}\left( \tau, k \right) + \frac{k^{2} \delta\left( \tau, k \right) \mathtt{c_s^2}\left( \tau, k \right)}{1 + w\left( \tau \right)} + k^{2} \Psi\left( \tau, k \right) - k^{2} \sigma\left( \tau, k \right) + \left( -1 + 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau, k \right) \\ u\left( \tau, k \right) &= \frac{\theta\left( \tau, k \right)}{k} \\ \mathtt{\dot{u}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} u\left( \tau, k \right) \\ \sigma\left( \tau, k \right) &= 0 \\ \mathtt{{\theta}interaction}\left( \tau, k \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(c)

\[ \begin{align*} \delta\left( \tau, k \right) &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.baryons — Function

baryons(g; recombination=true, name = :b, kwargs...)

Create a particle species for baryons in the spacetime with metric g.

b = SymBoltz.baryons(g)

Equations

equations(b)

\[ \begin{align*} \mathtt{c_s^2}\left( \tau, k \right) &= \mathtt{rec.c_s^2}\left( \tau \right) \\ w\left( \tau \right) &= 0 \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= 0.11937 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( \tau \right) &= 8.3776 \rho\left( \tau \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \delta\left( \tau, k \right) &= \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau, k \right) - 3 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau, k \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \theta\left( \tau, k \right) &= \mathtt{{\theta}interaction}\left( \tau, k \right) + \frac{k^{2} \delta\left( \tau, k \right) \mathtt{c_s^2}\left( \tau, k \right)}{1 + w\left( \tau \right)} + k^{2} \Psi\left( \tau, k \right) - k^{2} \sigma\left( \tau, k \right) + \left( -1 + 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau, k \right) \\ u\left( \tau, k \right) &= \frac{\theta\left( \tau, k \right)}{k} \\ \mathtt{\dot{u}}\left( \tau, k \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} u\left( \tau, k \right) \\ \sigma\left( \tau, k \right) &= 0 \\ \mathtt{rec.fHe} &= \frac{\mathtt{rec.Yp}}{3.9708 \left( 1 - \mathtt{rec.Yp} \right)} \\ \mathtt{rec.nH}\left( \tau \right) &= 5.9743 \cdot 10^{26} \left( 1 - \mathtt{rec.Yp} \right) \mathtt{rec.{\rho}b}\left( \tau \right) \\ \mathtt{rec.nHe}\left( \tau \right) &= \mathtt{rec.fHe} \mathtt{rec.nH}\left( \tau \right) \\ \mathtt{rec.DTb}\left( \tau \right) &= \frac{ - 0.00015165 \left( \mathtt{rec.T\gamma}\left( \tau \right) \right)^{4} a\left( \tau \right) \mathtt{rec.{\Delta}T}\left( \tau \right) \mathtt{rec.Xe}\left( \tau \right)}{h \left( 1 + \mathtt{rec.fHe} + \mathtt{rec.Xe}\left( \tau \right) \right)} - 2 \mathtt{rec.Tb}\left( \tau \right) \mathscr{E}\left( \tau \right) \\ \mathtt{rec.DT\gamma}\left( \tau \right) &= \frac{\mathrm{d} \mathtt{rec.T\gamma}\left( \tau \right)}{\mathrm{d}\tau} \\ \frac{\mathrm{d} \mathtt{rec.{\Delta}T}\left( \tau \right)}{\mathrm{d}\tau} &= \mathtt{rec.DTb}\left( \tau \right) - \mathtt{rec.DT\gamma}\left( \tau \right) \\ \mathtt{rec.Tb}\left( \tau \right) &= \mathtt{rec.{\Delta}T}\left( \tau \right) + \mathtt{rec.T\gamma}\left( \tau \right) \\ \mathtt{rec.{\beta}b}\left( \tau \right) &= \frac{1}{1.3806 \cdot 10^{-23} \mathtt{rec.Tb}\left( \tau \right)} \\ \mathtt{rec.{\lambda}e}\left( \tau \right) &= \frac{6.6261 \cdot 10^{-34}}{\sqrt{\frac{5.7236 \cdot 10^{-30}}{\mathtt{rec.{\beta}b}\left( \tau \right)}}} \\ \mathtt{rec.\mu}\left( \tau \right) &= \frac{1.6738 \cdot 10^{-27}}{1 - 0.74816 \mathtt{rec.Yp} + \left( 1 - \mathtt{rec.Yp} \right) \mathtt{rec.Xe}\left( \tau \right)} \\ \mathtt{rec.c_s^2}\left( \tau \right) &= \frac{1.3806 \cdot 10^{-23} \left( \mathtt{rec.Tb}\left( \tau \right) + \frac{ - \frac{\mathrm{d} \mathtt{rec.Tb}\left( \tau \right)}{\mathrm{d}\tau}}{3 \mathscr{E}\left( \tau \right)} \right)}{8.9876 \cdot 10^{16} \mathtt{rec.\mu}\left( \tau \right)} \\ \mathtt{rec.{\alpha}H}\left( \tau \right) &= \frac{4.9123 \cdot 10^{-19}}{0.0034166 \left( \mathtt{rec.Tb}\left( \tau \right) \right)^{0.6166} \left( 1 + 0.0050847 \left( \mathtt{rec.Tb}\left( \tau \right) \right)^{0.53} \right)} \\ \mathtt{rec.{\beta}H}\left( \tau \right) &= \frac{\mathtt{rec.{\alpha}H}\left( \tau \right) e^{ - 5.4468 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)}}{\left( \mathtt{rec.{\lambda}e}\left( \tau \right) \right)^{3}} \\ \mathtt{rec.KH0}\left( \tau \right) &= \frac{1.7966 \cdot 10^{-21}}{25.133 H\left( \tau \right)} \\ \mathtt{rec.KH1}\left( \tau \right) &= \left( 0.079 e^{ - 9.1827 \left( 6.73 + \log\left( a\left( \tau \right) \right) \right)^{2}} - 0.14 e^{ - 30.864 \left( 7.28 + \log\left( a\left( \tau \right) \right) \right)^{2}} \right) \mathtt{rec.KH0}\left( \tau \right) \\ \mathtt{rec.KH}\left( \tau \right) &= \mathtt{rec.KH1}\left( \tau \right) + \mathtt{rec.KH0}\left( \tau \right) \\ \mathtt{rec.CH}\left( \tau \right) &= \frac{1 + 8.2246 \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right) \mathtt{rec.KH}\left( \tau \right) \mathtt{rec.nH}\left( \tau \right)}{1 + \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right) \mathtt{rec.KH}\left( \tau \right) \left( 8.2246 + \mathtt{rec.{\beta}H}\left( \tau \right) \right) \mathtt{rec.nH}\left( \tau \right)} \\ \frac{\mathrm{d} \mathtt{rec.XH^+}\left( \tau \right)}{\mathrm{d}\tau} &= \frac{ - a\left( \tau \right) \left( \mathtt{rec.XH^+}\left( \tau \right) \mathtt{rec.{\alpha}H}\left( \tau \right) \mathtt{rec.ne}\left( \tau \right) - \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right) \mathtt{rec.{\beta}H}\left( \tau \right) e^{ - 1.634 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( \tau \right)} \right) \mathtt{rec.CH}\left( \tau \right)}{3.2408 \cdot 10^{-18} h} \\ \mathtt{rec.{\alpha}He}\left( \tau \right) &= \frac{1.803 \cdot 10^{-17}}{\left( 1 + \sqrt{7.6913 \cdot 10^{-6} \mathtt{rec.Tb}\left( \tau \right)} \right)^{1.711} \left( 1 + \sqrt{0.33333 \mathtt{rec.Tb}\left( \tau \right)} \right)^{0.289} \sqrt{0.33333 \mathtt{rec.Tb}\left( \tau \right)}} \\ \mathtt{rec.{\beta}He}\left( \tau \right) &= \frac{4 e^{ - 6.3633 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)} \mathtt{rec.{\alpha}He}\left( \tau \right)}{\left( \mathtt{rec.{\lambda}e}\left( \tau \right) \right)^{3}} \\ \mathtt{rec.KHe0^{- 1}}\left( \tau \right) &= 1.2597 \cdot 10^{23} H\left( \tau \right) \\ \mathtt{rec.KHe1^{- 1}}\left( \tau \right) &= - e^{\frac{ - 5.3949 \cdot 10^{9} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right)}{\mathtt{rec.KHe0^{- 1}}\left( \tau \right)}} \mathtt{rec.KHe0^{- 1}}\left( \tau \right) \\ \mathtt{rec.{\gamma}2Ps}\left( \tau \right) &= \left|\frac{4.8487 \cdot 10^{26} \mathtt{rec.fHe} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right)}{4.8748 \cdot 10^{26} \sqrt{\frac{6.2832}{5.9736 \cdot 10^{-10} \mathtt{rec.{\beta}b}\left( \tau \right)}} \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right)}\right| \\ \mathtt{rec.KHe2^{- 1}}\left( \tau \right) &= \frac{5.3949 \cdot 10^{9} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right)}{1 + 0.36 \left( \mathtt{rec.{\gamma}2Ps}\left( \tau \right) \right)^{0.86}} \\ \mathtt{rec.KHe}\left( \tau \right) &= \frac{1}{\mathtt{rec.KHe2^{- 1}}\left( \tau \right) + \mathtt{rec.KHe0^{- 1}}\left( \tau \right) + \mathtt{rec.KHe1^{- 1}}\left( \tau \right)} \\ \mathtt{rec.CHe}\left( \tau \right) &= \frac{e^{ - 9.6491 \cdot 10^{-20} \mathtt{rec.{\beta}b}\left( \tau \right)} + 51.3 \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right) \mathtt{rec.KHe}\left( \tau \right)}{e^{ - 9.6491 \cdot 10^{-20} \mathtt{rec.{\beta}b}\left( \tau \right)} + \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \left( 51.3 + \mathtt{rec.{\beta}He}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right) \mathtt{rec.KHe}\left( \tau \right)} \\ \mathtt{rec.DXHe^+\_singlet}\left( \tau \right) &= \frac{ - a\left( \tau \right) \left( - \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.{\beta}He}\left( \tau \right) e^{ - 3.303 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( \tau \right)} + \mathtt{rec.XHe^+}\left( \tau \right) \mathtt{rec.{\alpha}He}\left( \tau \right) \mathtt{rec.ne}\left( \tau \right) \right) \mathtt{rec.CHe}\left( \tau \right)}{3.2408 \cdot 10^{-18} h} \\ \mathtt{rec.{\alpha}He3}\left( \tau \right) &= \frac{4.9431 \cdot 10^{-17}}{\left( 1 + \sqrt{7.6913 \cdot 10^{-6} \mathtt{rec.Tb}\left( \tau \right)} \right)^{1.761} \left( 1 + \sqrt{0.33333 \mathtt{rec.Tb}\left( \tau \right)} \right)^{0.239} \sqrt{0.33333 \mathtt{rec.Tb}\left( \tau \right)}} \\ \mathtt{rec.{\beta}He3}\left( \tau \right) &= \frac{1.3333 e^{ - 7.6388 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)} \mathtt{rec.{\alpha}He3}\left( \tau \right)}{\left( \mathtt{rec.{\lambda}e}\left( \tau \right) \right)^{3}} \\ \mathtt{rec.{\tau}He3}\left( \tau \right) &= \frac{1.102 \cdot 10^{-19} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right)}{25.133 H\left( \tau \right)} \\ \mathtt{rec.pHe3}\left( \tau \right) &= \frac{1 - e^{ - \mathtt{rec.{\tau}He3}\left( \tau \right)}}{\mathtt{rec.{\tau}He3}\left( \tau \right)} \\ \mathtt{rec.{\gamma}2Pt}\left( \tau \right) &= \left|\frac{1.8633 \cdot 10^{-12} \mathtt{rec.fHe} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right)}{1.8917 \cdot 10^{-5} \sqrt{\frac{6.2832}{5.9736 \cdot 10^{-10} \mathtt{rec.{\beta}b}\left( \tau \right)}} \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right)}\right| \\ \mathtt{rec.CHe3}\left( \tau \right) &= \frac{1 \cdot 10^{-10} + 177.58 \left( \mathtt{rec.pHe3}\left( \tau \right) + \frac{\frac{1}{3}}{1 + 0.66 \left( \mathtt{rec.{\gamma}2Pt}\left( \tau \right) \right)^{0.9}} \right) e^{ - 1.8337 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)}}{1 \cdot 10^{-10} + \mathtt{rec.{\beta}He3}\left( \tau \right) + 177.58 \left( \mathtt{rec.pHe3}\left( \tau \right) + \frac{\frac{1}{3}}{1 + 0.66 \left( \mathtt{rec.{\gamma}2Pt}\left( \tau \right) \right)^{0.9}} \right) e^{ - 1.8337 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)}} \\ \mathtt{rec.DXHe^+\_triplet}\left( \tau \right) &= \frac{ - a\left( \tau \right) \left( - 3 \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) e^{ - 3.1755 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( \tau \right)} \mathtt{rec.{\beta}He3}\left( \tau \right) + \mathtt{rec.XHe^+}\left( \tau \right) \mathtt{rec.ne}\left( \tau \right) \mathtt{rec.{\alpha}He3}\left( \tau \right) \right) \mathtt{rec.CHe3}\left( \tau \right)}{3.2408 \cdot 10^{-18} h} \\ \frac{\mathrm{d} \mathtt{rec.XHe^+}\left( \tau \right)}{\mathrm{d}\tau} &= \mathtt{rec.DXHe^+\_triplet}\left( \tau \right) + \mathtt{rec.DXHe^+\_singlet}\left( \tau \right) \\ \mathtt{rec.RHe^+}\left( \tau \right) &= \frac{e^{ - 8.7187 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( \tau \right)}}{\left( \mathtt{rec.{\lambda}e}\left( \tau \right) \right)^{3} \mathtt{rec.nH}\left( \tau \right)} \\ \mathtt{rec.XHe^{+ +}}\left( \tau \right) &= \frac{2 \mathtt{rec.fHe} \mathtt{rec.RHe^+}\left( \tau \right)}{\left( 1 + \mathtt{rec.fHe} + \mathtt{rec.RHe^+}\left( \tau \right) \right) \left( 1 + \sqrt{1 + \frac{4 \mathtt{rec.fHe} \mathtt{rec.RHe^+}\left( \tau \right)}{\left( 1 + \mathtt{rec.fHe} + \mathtt{rec.RHe^+}\left( \tau \right) \right)^{2}}} \right)} \\ \mathtt{rec.re1Xe}\left( \tau \right) &= 0.5 \left( 1 + \mathtt{rec.fHe} + \left( 1 + \mathtt{rec.fHe} \right) \tanh\left( \frac{\left( 1 + \mathtt{rec.re1z} \right)^{1.5} - \left( 1 + z\left( \tau \right) \right)^{1.5}}{0.75 \left( 1 + \mathtt{rec.re1z} \right)^{0.5}} \right) \right) \\ \mathtt{rec.re2Xe}\left( \tau \right) &= 0.5 \left( \mathtt{rec.fHe} + \mathtt{rec.fHe} \tanh\left( \frac{\left( 1 + \mathtt{rec.re2z} \right)^{1.5} - \left( 1 + z\left( \tau \right) \right)^{1.5}}{0.75 \left( 1 + \mathtt{rec.re2z} \right)^{0.5}} \right) \right) \\ \mathtt{rec.Xe}\left( \tau \right) &= \mathtt{rec.XH^+}\left( \tau \right) + \mathtt{rec.re1Xe}\left( \tau \right) + \mathtt{rec.re2Xe}\left( \tau \right) + \mathtt{rec.XHe^{+ +}}\left( \tau \right) + \mathtt{rec.fHe} \mathtt{rec.XHe^+}\left( \tau \right) \\ \mathtt{rec.ne}\left( \tau \right) &= \mathtt{rec.Xe}\left( \tau \right) \mathtt{rec.nH}\left( \tau \right) \\ \frac{\mathrm{d} \mathtt{rec.\_\kappa}\left( \tau \right)}{\mathrm{d}\tau} &= \frac{ - 1.9944 \cdot 10^{-20} a\left( \tau \right) \mathtt{rec.ne}\left( \tau \right)}{3.2408 \cdot 10^{-18} h} \\ \mathtt{rec.\dot{\kappa}}\left( \tau \right) &= \frac{\mathrm{d} \mathtt{rec.\_\kappa}\left( \tau \right)}{\mathrm{d}\tau} \\ \mathtt{rec.\kappa}\left( \tau \right) &= - \mathtt{rec.{\kappa}0} + \mathtt{rec.\_\kappa}\left( \tau \right) \\ \mathtt{rec.I}\left( \tau \right) &= e^{ - \mathtt{rec.\kappa}\left( \tau \right)} \\ \mathtt{rec.v}\left( \tau \right) &= - \frac{\mathrm{d} \mathtt{rec.\kappa}\left( \tau \right)}{\mathrm{d}\tau} e^{ - \mathtt{rec.\kappa}\left( \tau \right)} \\ \mathtt{rec.\dot{v}}\left( \tau \right) &= \frac{\mathrm{d} \mathtt{rec.v}\left( \tau \right)}{\mathrm{d}\tau} \end{align*} \]

Initialization equations

initialization_equations(b)

\[ \begin{align*} \delta\left( \tau, k \right) &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \end{align*} \]

Neutrinos

SymBoltz.massless_neutrinos — Function

massless_neutrinos(g; lmax = 6, name = :ν, kwargs...)

Create a particle species for massless neutrinos in the spacetime with metric g.

ν = SymBoltz.massless_neutrinos(g; lmax = 6)

Equations

equations(ν)

\[ \begin{align*} \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{F0}\left( \tau, k \right) &= 4 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) - F\_{1}\left( \tau, k \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{1}\left( \tau, k \right) &= \frac{1}{3} \left( - 2 F\_{2}\left( \tau, k \right) + 4 \Psi\left( \tau, k \right) + \mathtt{F0}\left( \tau, k \right) \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{2}\left( \tau, k \right) &= \frac{1}{5} \left( 2 F\_{1}\left( \tau, k \right) - 3 F\_{3}\left( \tau, k \right) \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{3}\left( \tau, k \right) &= \frac{1}{7} \left( 3 F\_{2}\left( \tau, k \right) - 4 F\_{4}\left( \tau, k \right) \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{4}\left( \tau, k \right) &= \frac{1}{9} \left( 4 F\_{3}\left( \tau, k \right) - 5 F\_{5}\left( \tau, k \right) \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{5}\left( \tau, k \right) &= \frac{1}{11} \left( 5 F\_{4}\left( \tau, k \right) - 6 F\_{6}\left( \tau, k \right) \right) k \\ \frac{\mathrm{d}}{\mathrm{d}\tau} F\_{6}\left( \tau, k \right) &= F\_{5}\left( \tau, k \right) k - 7 F\_{6}\left( \tau, k \right) \mathscr{E}\left( \tau \right) \\ \delta\left( \tau, k \right) &= \mathtt{F0}\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{3}{4} F\_{1}\left( \tau, k \right) k \\ \sigma\left( \tau, k \right) &= \frac{1}{2} F\_{2}\left( \tau, k \right) \\ \mathtt{c_s^2}\left( \tau, k \right) &= \frac{1}{3} \\ T\left( \tau \right) &= \frac{\mathtt{T{_0}}}{a\left( \tau \right)} \\ w\left( \tau \right) &= \frac{1}{3} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= 0.11937 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( \tau \right) &= 8.3776 \rho\left( \tau \right) \end{align*} \]

Initialization equations

initialization_equations(ν)

\[ \begin{align*} \delta\left( \tau, k \right) &= - 2 \Psi\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \\ \sigma\left( \tau, k \right) &= \frac{1}{15} \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} \Psi\left( \tau, k \right) \\ F\_{3}\left( \tau, k \right) &= \frac{\frac{3}{7} F\_{2}\left( \tau, k \right) k}{\mathscr{E}\left( \tau \right)} \\ F\_{4}\left( \tau, k \right) &= \frac{\frac{4}{9} F\_{3}\left( \tau, k \right) k}{\mathscr{E}\left( \tau \right)} \\ F\_{5}\left( \tau, k \right) &= \frac{\frac{5}{11} F\_{4}\left( \tau, k \right) k}{\mathscr{E}\left( \tau \right)} \\ F\_{6}\left( \tau, k \right) &= \frac{\frac{6}{13} F\_{5}\left( \tau, k \right) k}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.massive_neutrinos — Function

massive_neutrinos(g; nx = 5, lmax = 4, name = :h, kwargs...)

Create a particle species for massive neutrinos in the spacetime with metric g.

h = SymBoltz.massive_neutrinos(g; nx=4, lmax = 5)

Equations

equations(h)

\[ \begin{align*} m &= 1.7827 \cdot 10^{-36} \mathtt{m\_eV} \\ \mathtt{y{_0}} &= \frac{8.9876 \cdot 10^{16} m}{1.3806 \cdot 10^{-23} \mathtt{T{_0}}} \\ \mathtt{I\rho{_0}} &= 0.62542 \sqrt{1.7586 + \mathtt{y{_0}}^{2}} + 0.98076 \sqrt{12.395 + \mathtt{y{_0}}^{2}} + 0.19286 \sqrt{48.934 + \mathtt{y{_0}}^{2}} + 0.0040482 \sqrt{157.36 + \mathtt{y{_0}}^{2}} \\ \mathtt{\Omega{_0}} &= \frac{9.7603 \cdot 10^{-16} \mathtt{T{_0}}^{4} \mathtt{I\rho{_0}}}{1.4143 \cdot 10^{-8} h^{2}} \\ T\left( \tau \right) &= \frac{\mathtt{T{_0}}}{a\left( \tau \right)} \\ y\left( \tau \right) &= \mathtt{y{_0}} a\left( \tau \right) \\ \mathtt{In}\left( \tau \right) &= 1.8031 \\ \mathtt{I\rho}\left( \tau \right) &= 0.62542 E\_{1}\left( \tau \right) + 0.98076 E\_{2}\left( \tau \right) + 0.19286 E\_{3}\left( \tau \right) + 0.0040482 E\_{4}\left( \tau \right) \\ \mathtt{IP}\left( \tau \right) &= \frac{12.157}{E\_{2}\left( \tau \right)} + \frac{9.4373}{E\_{3}\left( \tau \right)} + \frac{0.63702}{E\_{4}\left( \tau \right)} + \frac{1.0999}{E\_{1}\left( \tau \right)} \\ \rho\left( \tau \right) &= \frac{1.165 \cdot 10^{-16} \left( T\left( \tau \right) \right)^{4} \mathtt{I\rho}\left( \tau \right)}{1.4143 \cdot 10^{-8} h^{2}} \\ P\left( \tau \right) &= \frac{3.8835 \cdot 10^{-17} \left( T\left( \tau \right) \right)^{4} \mathtt{IP}\left( \tau \right)}{1.4143 \cdot 10^{-8} h^{2}} \\ w\left( \tau \right) &= \frac{P\left( \tau \right)}{\rho\left( \tau \right)} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ \mathtt{I\delta\rho}\left( \tau, k \right) &= 0.62542 E\_{1}\left( \tau \right) \mathtt{{\psi}0}\_{1}\left( \tau, k \right) + 0.98076 E\_{2}\left( \tau \right) \mathtt{{\psi}0}\_{2}\left( \tau, k \right) + 0.19286 E\_{3}\left( \tau \right) \mathtt{{\psi}0}\_{3}\left( \tau, k \right) + 0.0040482 E\_{4}\left( \tau \right) \mathtt{{\psi}0}\_{4}\left( \tau, k \right) \\ \delta\left( \tau, k \right) &= \frac{\mathtt{I\delta\rho}\left( \tau, k \right)}{\mathtt{I\rho}\left( \tau \right)} \\ u\left( \tau, k \right) &= \frac{0.82938 \psi\_{1,1}\left( \tau, k \right) + 3.4529 \psi\_{2,1}\left( \tau, k \right) + 1.3491 \psi\_{3,1}\left( \tau, k \right) + 0.050782 \psi\_{4,1}\left( \tau, k \right)}{\mathtt{I\rho}\left( \tau \right) + \frac{1}{3} \mathtt{IP}\left( \tau \right)} \\ \theta\left( \tau, k \right) &= k u\left( \tau, k \right) \\ \sigma\left( \tau, k \right) &= \frac{\frac{2}{3} \left( \frac{9.4373 \psi\_{3,2}\left( \tau, k \right)}{E\_{3}\left( \tau \right)} + \frac{12.157 \psi\_{2,2}\left( \tau, k \right)}{E\_{2}\left( \tau \right)} + \frac{1.0999 \psi\_{1,2}\left( \tau, k \right)}{E\_{1}\left( \tau \right)} + \frac{0.63702 \psi\_{4,2}\left( \tau, k \right)}{E\_{4}\left( \tau \right)} \right)}{\mathtt{I\rho}\left( \tau \right) + \frac{1}{3} \mathtt{IP}\left( \tau \right)} \\ \mathtt{c_s^2}\left( \tau, k \right) &= \frac{\frac{0.63702 \mathtt{{\psi}0}\_{4}\left( \tau, k \right)}{E\_{4}\left( \tau \right)} + \frac{9.4373 \mathtt{{\psi}0}\_{3}\left( \tau, k \right)}{E\_{3}\left( \tau \right)} + \frac{1.0999 \mathtt{{\psi}0}\_{1}\left( \tau, k \right)}{E\_{1}\left( \tau \right)} + \frac{12.157 \mathtt{{\psi}0}\_{2}\left( \tau, k \right)}{E\_{2}\left( \tau \right)}}{\mathtt{I\delta\rho}\left( \tau, k \right)} \\ E\_{1}\left( \tau \right) &= \sqrt{1.7586 + \left( y\left( \tau \right) \right)^{2}} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{{\psi}0}\_{1}\left( \tau, k \right) &= \frac{ - 1.3261 k \psi\_{1,1}\left( \tau, k \right)}{E\_{1}\left( \tau \right)} + 1.0479 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{1,1}\left( \tau, k \right) &= \frac{0.44204 k \left( - 2 \psi\_{1,2}\left( \tau, k \right) + \mathtt{{\psi}0}\_{1}\left( \tau, k \right) \right)}{E\_{1}\left( \tau \right)} + 0.2634 E\_{1}\left( \tau \right) k \Psi\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{1,2}\left( \tau, k \right) &= \frac{0.26522 k \left( 2 \psi\_{1,1}\left( \tau, k \right) - 3 \psi\_{1,3}\left( \tau, k \right) \right)}{E\_{1}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{1,3}\left( \tau, k \right) &= \frac{0.18945 k \left( 3 \psi\_{1,2}\left( \tau, k \right) - 4 \psi\_{1,4}\left( \tau, k \right) \right)}{E\_{1}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{1,4}\left( \tau, k \right) &= \frac{0.14735 k \left( 4 \psi\_{1,3}\left( \tau, k \right) - 5 \psi\_{1,5}\left( \tau, k \right) \right)}{E\_{1}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{1,5}\left( \tau, k \right) &= \frac{0.12056 k \left( - 6 \left( \frac{8.2949 E\_{1}\left( \tau \right) \mathscr{E}\left( \tau \right) \psi\_{1,5}\left( \tau, k \right)}{k} - \psi\_{1,4}\left( \tau, k \right) \right) + 5 \psi\_{1,4}\left( \tau, k \right) \right)}{E\_{1}\left( \tau \right)} \\ E\_{2}\left( \tau \right) &= \sqrt{12.395 + \left( y\left( \tau \right) \right)^{2}} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{{\psi}0}\_{2}\left( \tau, k \right) &= \frac{ - 3.5207 k \psi\_{2,1}\left( \tau, k \right)}{E\_{2}\left( \tau \right)} + 3.4195 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{2,1}\left( \tau, k \right) &= \frac{1.1736 k \left( - 2 \psi\_{2,2}\left( \tau, k \right) + \mathtt{{\psi}0}\_{2}\left( \tau, k \right) \right)}{E\_{2}\left( \tau \right)} + 0.32376 E\_{2}\left( \tau \right) k \Psi\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{2,2}\left( \tau, k \right) &= \frac{0.70413 k \left( 2 \psi\_{2,1}\left( \tau, k \right) - 3 \psi\_{2,3}\left( \tau, k \right) \right)}{E\_{2}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{2,3}\left( \tau, k \right) &= \frac{0.50295 k \left( 3 \psi\_{2,2}\left( \tau, k \right) - 4 \psi\_{2,4}\left( \tau, k \right) \right)}{E\_{2}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{2,4}\left( \tau, k \right) &= \frac{0.39119 k \left( 4 \psi\_{2,3}\left( \tau, k \right) - 5 \psi\_{2,5}\left( \tau, k \right) \right)}{E\_{2}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{2,5}\left( \tau, k \right) &= \frac{0.32006 k \left( - 6 \left( \frac{3.1244 E\_{2}\left( \tau \right) \mathscr{E}\left( \tau \right) \psi\_{2,5}\left( \tau, k \right)}{k} - \psi\_{2,4}\left( \tau, k \right) \right) + 5 \psi\_{2,4}\left( \tau, k \right) \right)}{E\_{2}\left( \tau \right)} \\ E\_{3}\left( \tau \right) &= \sqrt{48.934 + \left( y\left( \tau \right) \right)^{2}} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{{\psi}0}\_{3}\left( \tau, k \right) &= \frac{ - 6.9953 k \psi\_{3,1}\left( \tau, k \right)}{E\_{3}\left( \tau \right)} + 6.9889 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{3,1}\left( \tau, k \right) &= \frac{2.3318 k \left( - 2 \psi\_{3,2}\left( \tau, k \right) + \mathtt{{\psi}0}\_{3}\left( \tau, k \right) \right)}{E\_{3}\left( \tau \right)} + 0.33303 E\_{3}\left( \tau \right) k \Psi\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{3,2}\left( \tau, k \right) &= \frac{1.3991 k \left( 2 \psi\_{3,1}\left( \tau, k \right) - 3 \psi\_{3,3}\left( \tau, k \right) \right)}{E\_{3}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{3,3}\left( \tau, k \right) &= \frac{0.99933 k \left( 3 \psi\_{3,2}\left( \tau, k \right) - 4 \psi\_{3,4}\left( \tau, k \right) \right)}{E\_{3}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{3,4}\left( \tau, k \right) &= \frac{0.77726 k \left( 4 \psi\_{3,3}\left( \tau, k \right) - 5 \psi\_{3,5}\left( \tau, k \right) \right)}{E\_{3}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{3,5}\left( \tau, k \right) &= \frac{0.63594 k \left( - 6 \left( \frac{1.5725 E\_{3}\left( \tau \right) \mathscr{E}\left( \tau \right) \psi\_{3,5}\left( \tau, k \right)}{k} - \psi\_{3,4}\left( \tau, k \right) \right) + 5 \psi\_{3,4}\left( \tau, k \right) \right)}{E\_{3}\left( \tau \right)} \\ E\_{4}\left( \tau \right) &= \sqrt{157.36 + \left( y\left( \tau \right) \right)^{2}} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \mathtt{{\psi}0}\_{4}\left( \tau, k \right) &= \frac{ - 12.544 k \psi\_{4,1}\left( \tau, k \right)}{E\_{4}\left( \tau \right)} + 12.544 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{4,1}\left( \tau, k \right) &= \frac{4.1814 k \left( - 2 \psi\_{4,2}\left( \tau, k \right) + \mathtt{{\psi}0}\_{4}\left( \tau, k \right) \right)}{E\_{4}\left( \tau \right)} + 0.33333 E\_{4}\left( \tau \right) k \Psi\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{4,2}\left( \tau, k \right) &= \frac{2.5089 k \left( 2 \psi\_{4,1}\left( \tau, k \right) - 3 \psi\_{4,3}\left( \tau, k \right) \right)}{E\_{4}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{4,3}\left( \tau, k \right) &= \frac{1.792 k \left( 3 \psi\_{4,2}\left( \tau, k \right) - 4 \psi\_{4,4}\left( \tau, k \right) \right)}{E\_{4}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{4,4}\left( \tau, k \right) &= \frac{1.3938 k \left( 4 \psi\_{4,3}\left( \tau, k \right) - 5 \psi\_{4,5}\left( \tau, k \right) \right)}{E\_{4}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \psi\_{4,5}\left( \tau, k \right) &= \frac{1.1404 k \left( - 6 \left( \frac{0.8769 E\_{4}\left( \tau \right) \mathscr{E}\left( \tau \right) \psi\_{4,5}\left( \tau, k \right)}{k} - \psi\_{4,4}\left( \tau, k \right) \right) + 5 \psi\_{4,4}\left( \tau, k \right) \right)}{E\_{4}\left( \tau \right)} \end{align*} \]

Initialization equations

initialization_equations(h)

\[ \begin{align*} \mathtt{{\psi}0}\_{1}\left( \tau, k \right) &= - 0.52395 \Psi\left( \tau, k \right) \\ \psi\_{1,1}\left( \tau, k \right) &= \frac{0.1317 E\_{1}\left( \tau \right) k \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \\ \psi\_{1,2}\left( \tau, k \right) &= 0.03493 \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} \Psi\left( \tau, k \right) \\ \psi\_{1,3}\left( \tau, k \right) &= 0 \\ \psi\_{1,4}\left( \tau, k \right) &= 0 \\ \psi\_{1,5}\left( \tau, k \right) &= 0 \\ \mathtt{{\psi}0}\_{2}\left( \tau, k \right) &= - 1.7098 \Psi\left( \tau, k \right) \\ \psi\_{2,1}\left( \tau, k \right) &= \frac{0.16188 E\_{2}\left( \tau \right) k \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \\ \psi\_{2,2}\left( \tau, k \right) &= 0.11398 \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} \Psi\left( \tau, k \right) \\ \psi\_{2,3}\left( \tau, k \right) &= 0 \\ \psi\_{2,4}\left( \tau, k \right) &= 0 \\ \psi\_{2,5}\left( \tau, k \right) &= 0 \\ \mathtt{{\psi}0}\_{3}\left( \tau, k \right) &= - 3.4945 \Psi\left( \tau, k \right) \\ \psi\_{3,1}\left( \tau, k \right) &= \frac{0.16651 E\_{3}\left( \tau \right) k \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \\ \psi\_{3,2}\left( \tau, k \right) &= 0.23296 \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} \Psi\left( \tau, k \right) \\ \psi\_{3,3}\left( \tau, k \right) &= 0 \\ \psi\_{3,4}\left( \tau, k \right) &= 0 \\ \psi\_{3,5}\left( \tau, k \right) &= 0 \\ \mathtt{{\psi}0}\_{4}\left( \tau, k \right) &= - 6.2721 \Psi\left( \tau, k \right) \\ \psi\_{4,1}\left( \tau, k \right) &= \frac{0.16667 E\_{4}\left( \tau \right) k \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \\ \psi\_{4,2}\left( \tau, k \right) &= 0.41814 \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} \Psi\left( \tau, k \right) \\ \psi\_{4,3}\left( \tau, k \right) &= 0 \\ \psi\_{4,4}\left( \tau, k \right) &= 0 \\ \psi\_{4,5}\left( \tau, k \right) &= 0 \end{align*} \]

Dark energy

SymBoltz.cosmological_constant — Function

cosmological_constant(g; name = :Λ, kwargs...)

Create a particle species for the cosmological constant (with equation of state w ~ -1) in the spacetime with metric g.

Λ = SymBoltz.cosmological_constant(g)

Equations

equations(Λ)

\[ \begin{align*} \delta\left( \tau, k \right) &= 0 \\ \theta\left( \tau, k \right) &= 0 \\ \sigma\left( \tau, k \right) &= 0 \\ \mathtt{c_s^2}\left( \tau, k \right) &= -1 \\ w\left( \tau \right) &= -1 \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= 0.11937 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( \tau \right) &= 8.3776 \rho\left( \tau \right) \end{align*} \]

Initialization equations

initialization_equations(Λ)

\[ \begin{align*} \end{align*} \]

SymBoltz.w0wa — Function

w0wa(g; kwargs...)

Create a particle species for the w₀-wₐ dark energy (CPL) parametrization in the spacetime with metric g.

X = SymBoltz.w0wa(g)

Equations

equations(X)

\[ \begin{align*} w\left( \tau \right) &= \mathtt{w0} + \mathtt{wa} \left( 1 - a\left( \tau \right) \right) \\ \dot{w}\left( \tau \right) &= \frac{\mathrm{d} w\left( \tau \right)}{\mathrm{d}\tau} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \frac{\mathrm{d} \rho\left( \tau \right)}{\mathrm{d}\tau} &= - 3 \left( 1 + w\left( \tau \right) \right) \rho\left( \tau \right) \mathscr{E}\left( \tau \right) \\ \mathtt{c_a^2}\left( \tau \right) &= w\left( \tau \right) + \frac{ - \dot{w}\left( \tau \right)}{3 \left( 1 + w\left( \tau \right) \right) \mathscr{E}\left( \tau \right)} \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \delta\left( \tau, k \right) &= \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau, k \right) - 3 \frac{\mathrm{d}}{\mathrm{d}\tau} \Phi\left( \tau, k \right) \right) - 3 \left( \mathtt{c_s^2} - w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau, k \right) - 9 \left( \frac{\mathscr{E}\left( \tau \right)}{k} \right)^{2} \left( \mathtt{c_s^2} - \mathtt{c_a^2}\left( \tau \right) \right) \left( 1 + w\left( \tau \right) \right) \theta\left( \tau, k \right) \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \theta\left( \tau, k \right) &= \frac{k^{2} \mathtt{c_s^2} \delta\left( \tau, k \right)}{1 + w\left( \tau \right)} - \left( 1 - 3 \mathtt{c_s^2} \right) \mathscr{E}\left( \tau \right) \theta\left( \tau, k \right) + k^{2} \Psi\left( \tau, k \right) - k^{2} \sigma\left( \tau, k \right) \\ \sigma\left( \tau, k \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(X)

\[ \begin{align*} \delta\left( \tau, k \right) &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau, k \right) \\ \theta\left( \tau, k \right) &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau, k \right)}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.quintessence — Function

quintessence(g, v; name = :ϕ, kwargs...)

Create a species for a quintessence scalar field with potential v in the spacetime with metric g.

@variables ϕ(τ) V(ϕ) V′(ϕ) V′′(ϕ)
Q = SymBoltz.quintessence(g, V, V′, V′′)

Equations

equations(Q)

\[ \begin{align*} \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} &= - 2 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \mathscr{E}\left( \tau \right) - \left( a\left( \tau \right) \right)^{2} \mathtt{V\prime}\left( \tau \right) \\ V\left( \tau \right) &= V\left( \phi\left( \tau \right) \right) \\ \mathtt{V\prime}\left( \tau \right) &= \mathtt{V\prime}\left( \phi\left( \tau \right) \right) \\ \mathtt{V\prime\prime}\left( \tau \right) &= \mathtt{V\prime\prime}\left( \phi\left( \tau \right) \right) \\ K\left( \tau \right) &= \frac{1}{2} \left( \frac{\frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{a\left( \tau \right)} \right)^{2} \\ \rho\left( \tau \right) &= V\left( \tau \right) + K\left( \tau \right) \\ P\left( \tau \right) &= - V\left( \tau \right) + K\left( \tau \right) \\ w\left( \tau \right) &= \frac{P\left( \tau \right)}{\rho\left( \tau \right)} \\ \mathtt{m^2}\left( \tau \right) &= \mathtt{V\prime\prime}\left( \tau \right) \\ \mathtt{{\epsilon}s}\left( \tau \right) &= \frac{\left( \frac{\mathtt{V\prime}\left( \tau \right)}{V\left( \tau \right)} \right)^{2}}{16 \pi} \\ \mathtt{{\eta}s}\left( \tau \right) &= \frac{\mathtt{V\prime\prime}\left( \tau \right)}{8 V\left( \tau \right) \pi} \\ \delta\left( \tau, k \right) &= 0 \\ \sigma\left( \tau, k \right) &= 0 \\ \mathtt{c_s^2}\left( \tau, k \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(Q)

\[ \begin{align*} \end{align*} \]

Curvature (effective)

SymBoltz.curvature — Function

curvature(g; name = :K, kwargs...)

Create a species that effectively accounts for curvature in the spacetime with metric g.

K = SymBoltz.curvature(g)

Equations

equations(K)

\[ \begin{align*} w\left( \tau \right) &= \frac{-1}{3} \\ \rho\left( \tau \right) &= \frac{3 \mathtt{\Omega_0}}{8 \left( a\left( \tau \right) \right)^{2} \pi} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \delta\left( \tau, k \right) &= 0 \\ \theta\left( \tau, k \right) &= 0 \\ \mathtt{c_s^2}\left( \tau \right) &= 0 \\ \sigma\left( \tau, k \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(K)

\[ \begin{align*} \end{align*} \]