Components

Spacetimes

SymBoltz.metric — Function

metric(; name = :g, kwargs...)

Create a symbolic component for the perturbed FLRW spacetime metric in the conformal Newtonian gauge with sign signature diag(-1, +1, +1, +1).

g = SymBoltz.metric()

Equations

equations(g)

\[ \begin{align*} z\left( \tau \right) &= -1 + \frac{1}{a\left( \tau \right)} \\ \dot{z}\left( \tau \right) &= \frac{\mathrm{d} z\left( \tau \right)}{\mathrm{d}\tau} \\ \mathscr{E}\left( \tau \right) &= \frac{\frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau}}{a\left( \tau \right)} \\ E\left( \tau \right) &= \frac{\mathscr{E}\left( \tau \right)}{a\left( \tau \right)} \\ \mathscr{H}\left( \tau \right) &= 3.2408 \cdot 10^{-18} h \mathscr{E}\left( \tau \right) \\ H\left( \tau \right) &= 3.2408 \cdot 10^{-18} h E\left( \tau \right) \\ \mathtt{g_{1 1}}\left( \tau \right) &= \left( a\left( \tau \right) \right)^{2} \left( -1 - 2 \Psi\left( \tau \right) \epsilon \right) \\ \mathtt{g_{2 2}}\left( \tau \right) &= \left( a\left( \tau \right) \right)^{2} \left( 1 - 2 \Phi\left( \tau \right) \epsilon \right) \end{align*} \]

Gravitational theories

SymBoltz.general_relativity — Function

general_relativity(g; name = :G, kwargs...)

Create a symbolic component for the general relativistic (GR) theory of gravity in the spacetime with the metric g.

G = SymBoltz.general_relativity(g)

Equations

equations(G)

\[ \begin{align*} \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} &= \left( a\left( \tau \right) \right)^{2} \sqrt{\frac{8}{3} \rho\left( \tau \right) \pi} \\ \mathtt{F_1}\left( \tau \right) &= 0 \\ \mathtt{F_2}\left( \tau \right) &= \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} + \frac{ - \left( \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} \right)^{2} \left( 1 + \frac{ - 3 P\left( \tau \right)}{\rho\left( \tau \right)} \right)}{2 a\left( \tau \right)} \\ \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{ - \frac{4}{3} \left( a\left( \tau \right) \right)^{2} \pi \mathtt{\delta\rho}\left( \tau \right)}{\mathscr{E}\left( \tau \right)} + \frac{ - k^{2} \Phi\left( \tau \right)}{3 \mathscr{E}\left( \tau \right)} - \Psi\left( \tau \right) \mathscr{E}\left( \tau \right) \right) \epsilon \\ k^{2} \left( \Phi\left( \tau \right) - \Psi\left( \tau \right) \right) \epsilon &= 12 \left( a\left( \tau \right) \right)^{2} \pi \Pi\left( \tau \right) \epsilon \end{align*} \]

Initialization equations

initialization_equations(G)

\[ \begin{align*} \end{align*} \]

SymBoltz.brans_dicke — Function

brans_dicke(g; name = :G, kwargs...)

Create a symbolic component for the Brans-Dicke (BD) theory of gravity in the spacetime with the metric g.

G = SymBoltz.brans_dicke(g)

Equations

equations(G)

\[ \begin{align*} \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} &= \frac{8 \left( a\left( \tau \right) \right)^{2} \left( - 3 P\left( \tau \right) + \rho\left( \tau \right) \right) \pi}{3 + 2 \omega} + \frac{ - 2 \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{a\left( \tau \right)} \\ G\left( \tau \right) &= \frac{4 + 2 \omega}{\phi\left( \tau \right) \left( 3 + 2 \omega \right)} \\ \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} &= \frac{ - \frac{1}{2} a\left( \tau \right) \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} + \sqrt{\frac{\frac{8}{3} \left( a\left( \tau \right) \right)^{4} \rho\left( \tau \right) \pi}{\phi\left( \tau \right)} + \left( \frac{\frac{1}{2} a\left( \tau \right) \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} \right)^{2} + \frac{1}{6} \left( \frac{a\left( \tau \right) \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} \right)^{2} \omega} \\ \Delta\left( \tau \right) &= \frac{ - \left( \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} \right)^{2}}{2 a\left( \tau \right)} + \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} + \frac{4 \left( a\left( \tau \right) \right)^{3} P\left( \tau \right) \pi}{\phi\left( \tau \right)} + \frac{\frac{1}{2} a\left( \tau \right) \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} + \frac{\frac{1}{2} \frac{\mathrm{d} a\left( \tau \right)}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} + \frac{1}{4} \left( \frac{\frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} \right)^{2} a\left( \tau \right) \omega \\ \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\left( 3 \mathscr{E}\left( \tau \right) \frac{\mathrm{d} \mathtt{\delta\phi}\left( \tau \right)}{\mathrm{d}\tau} + \left( k^{2} + 3 \left( \mathscr{E}\left( \tau \right) \right)^{2} \right) \mathtt{\delta\phi}\left( \tau \right) - 2 k^{2} \Phi\left( \tau \right) \phi\left( \tau \right) + \frac{1}{2} \left( - 2 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \frac{\mathrm{d} \mathtt{\delta\phi}\left( \tau \right)}{\mathrm{d}\tau} + \left( \frac{\frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{\phi\left( \tau \right)} \right)^{2} \mathtt{\delta\phi}\left( \tau \right) \right) \omega - 8 \left( a\left( \tau \right) \right)^{2} \left( \mathtt{\delta\rho}\left( \tau \right) + 2 \rho\left( \tau \right) \Psi\left( \tau \right) \right) \pi \right) \epsilon}{3 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} + 6 \mathscr{E}\left( \tau \right) \phi\left( \tau \right)} \\ \Psi\left( \tau \right) \epsilon &= \left( \frac{ - 12 \left( a\left( \tau \right) \right)^{2} \pi \Pi\left( \tau \right)}{k^{2} \phi\left( \tau \right)} + \Phi\left( \tau \right) + \frac{ - \mathtt{\delta\phi}\left( \tau \right)}{\phi\left( \tau \right)} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \mathtt{\delta\phi}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{ - 8 \left( a\left( \tau \right) \right)^{2} \left( 3 \mathtt{{\delta}P}\left( \tau \right) - \mathtt{\delta\rho}\left( \tau \right) \right) \pi}{3 + 2 \omega} + 2 \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \Psi\left( \tau \right) + \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \left( 3 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} + \frac{\mathrm{d} \Psi\left( \tau \right)}{\mathrm{d}\tau} \right) - 2 \mathscr{E}\left( \tau \right) \frac{\mathrm{d} \mathtt{\delta\phi}\left( \tau \right)}{\mathrm{d}\tau} - k^{2} \mathtt{\delta\phi}\left( \tau \right) + 4 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \Psi\left( \tau \right) \mathscr{E}\left( \tau \right) \right) \epsilon \end{align*} \]

Initialization equations

initialization_equations(G)

\[ \begin{align*} \mathtt{\delta\phi}\left( \tau \right) \epsilon &= 0 \\ \frac{\mathrm{d} \mathtt{\delta\phi}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= 0 \end{align*} \]

Species

Generic

SymBoltz.species_constant_eos — Function

species_constant_eos(g, w, cₛ² = w, ẇ = 0, _σ = 0; θinteract = false, kwargs...)

Create a symbolic component for a particle species with equation of state w ~ P/ρ in the spacetime with the metric g.

using SymBoltz: τ
@variables w(τ)
s = SymBoltz.species_constant_eos(g, w)

Equations

equations(s)

\[ \begin{align*} w\left( \tau \right) &= w\left( \tau \right) \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ \frac{\mathrm{d} \delta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau \right) - 3 \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{k^{2} \mathtt{c_s^2}\left( \tau \right) \delta\left( \tau \right)}{1 + w\left( \tau \right)} + \mathtt{{\theta}interaction}\left( \tau \right) - k^{2} \sigma\left( \tau \right) + k^{2} \Psi\left( \tau \right) - \left( 1 - 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau \right) \right) \epsilon \\ u\left( \tau \right) \epsilon &= \frac{\theta\left( \tau \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( \tau \right) \epsilon &= \frac{\mathrm{d} u\left( \tau \right)}{\mathrm{d}\tau} \epsilon \\ \Delta\left( \tau \right) \epsilon &= \left( \frac{3 \left( 1 + w\left( \tau \right) \right) \mathscr{E}\left( \tau \right)}{\theta\left( \tau \right)} + \delta\left( \tau \right) \right) \epsilon \\ \sigma\left( \tau \right) \epsilon &= 0 \\ \mathtt{{\theta}interaction}\left( \tau \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(s)

\[ \begin{align*} \delta\left( \tau \right) \epsilon &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right)} \end{align*} \]

Relativistic

SymBoltz.radiation — Function

radiation(g; name = :r, kwargs...)

Create a particle species for radiation (with equation of state w ~ 1/3) in the spacetime with metric g.

r = SymBoltz.radiation(g)

Equations

equations(r)

\[ \begin{align*} T\left( \tau \right) &= \frac{\mathtt{T{_0}}}{a\left( \tau \right)} \\ w\left( \tau \right) &= \frac{1}{3} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ \frac{\mathrm{d} \delta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau \right) - 3 \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{k^{2} \mathtt{c_s^2}\left( \tau \right) \delta\left( \tau \right)}{1 + w\left( \tau \right)} + \mathtt{{\theta}interaction}\left( \tau \right) - k^{2} \sigma\left( \tau \right) + k^{2} \Psi\left( \tau \right) - \left( 1 - 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau \right) \right) \epsilon \\ u\left( \tau \right) \epsilon &= \frac{\theta\left( \tau \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( \tau \right) \epsilon &= \frac{\mathrm{d} u\left( \tau \right)}{\mathrm{d}\tau} \epsilon \\ \Delta\left( \tau \right) \epsilon &= \left( \frac{3 \left( 1 + w\left( \tau \right) \right) \mathscr{E}\left( \tau \right)}{\theta\left( \tau \right)} + \delta\left( \tau \right) \right) \epsilon \\ \sigma\left( \tau \right) \epsilon &= 0 \\ \mathtt{{\theta}interaction}\left( \tau \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(r)

\[ \begin{align*} \delta\left( \tau \right) \epsilon &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.photons — Function

photons(g; polarization = true, lmax = 6, name = :γ, kwargs...)

Create a particle species for photons in the spacetime with metric g.

γ = SymBoltz.photons(g; lmax = 6)

Equations

equations(γ)

\[ \begin{align*} \frac{\mathrm{d} F\_{0}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( 4 \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} - F\_{1}\left( \tau \right) k \right) \epsilon \\ \frac{\mathrm{d} F\_{1}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{ - \frac{4}{3} \mathtt{\dot{\kappa}}\left( \tau \right) \left( - \theta\left( \tau \right) + \mathtt{{\theta}b}\left( \tau \right) \right)}{k} + \frac{1}{3} \left( F\_{0}\left( \tau \right) - 2 F\_{2}\left( \tau \right) + 4 \Psi\left( \tau \right) \right) k \right) \epsilon \\ \frac{\mathrm{d} F\_{2}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{5} \left( 2 F\_{1}\left( \tau \right) - 3 F\_{3}\left( \tau \right) \right) k + \frac{9}{10} F\_{2}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) - \frac{1}{10} \left( G\_{0}\left( \tau \right) + G\_{2}\left( \tau \right) \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} F\_{3}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{7} \left( 3 F\_{2}\left( \tau \right) - 4 F\_{4}\left( \tau \right) \right) k + F\_{3}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} F\_{4}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{9} \left( 4 F\_{3}\left( \tau \right) - 5 F\_{5}\left( \tau \right) \right) k + F\_{4}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} F\_{5}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{11} \left( 5 F\_{4}\left( \tau \right) - 6 F\_{6}\left( \tau \right) \right) k + F\_{5}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} F\_{6}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( F\_{5}\left( \tau \right) k + F\_{6}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) - 7 F\_{6}\left( \tau \right) \mathscr{E}\left( \tau \right) \right) \epsilon \\ \delta\left( \tau \right) \epsilon &= F\_{0}\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{3}{4} F\_{1}\left( \tau \right) k \epsilon \\ \sigma\left( \tau \right) \epsilon &= \frac{1}{2} F\_{2}\left( \tau \right) \epsilon \\ \Pi\left( \tau \right) \epsilon &= \left( F\_{2}\left( \tau \right) + G\_{0}\left( \tau \right) + G\_{2}\left( \tau \right) \right) \epsilon \\ \mathtt{\dot{\Pi}}\left( \tau \right) \epsilon &= \frac{\mathrm{d} \Pi\left( \tau \right)}{\mathrm{d}\tau} \epsilon \\ \mathtt{c_s^2}\left( \tau \right) \epsilon &= \frac{1}{3} \epsilon \\ \Theta\_{0}\left( \tau \right) \epsilon &= \frac{1}{4} F\_{0}\left( \tau \right) \epsilon \\ \Theta\_{1}\left( \tau \right) \epsilon &= \frac{1}{4} F\_{1}\left( \tau \right) \epsilon \\ \Theta\_{2}\left( \tau \right) \epsilon &= \frac{1}{4} F\_{2}\left( \tau \right) \epsilon \\ \Theta\_{3}\left( \tau \right) \epsilon &= \frac{1}{4} F\_{3}\left( \tau \right) \epsilon \\ \Theta\_{4}\left( \tau \right) \epsilon &= \frac{1}{4} F\_{4}\left( \tau \right) \epsilon \\ \Theta\_{5}\left( \tau \right) \epsilon &= \frac{1}{4} F\_{5}\left( \tau \right) \epsilon \\ \Theta\_{6}\left( \tau \right) \epsilon &= \frac{1}{4} F\_{6}\left( \tau \right) \epsilon \\ \frac{\mathrm{d} G\_{0}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \left( G\_{0}\left( \tau \right) - \frac{1}{2} \Pi\left( \tau \right) \right) \mathtt{\dot{\kappa}}\left( \tau \right) - G\_{1}\left( \tau \right) k \right) \epsilon \\ \frac{\mathrm{d} G\_{1}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{3} \left( G\_{0}\left( \tau \right) - 2 G\_{2}\left( \tau \right) \right) k + G\_{1}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} G\_{2}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{5} \left( 2 G\_{1}\left( \tau \right) - 3 G\_{3}\left( \tau \right) \right) k + \left( G\_{2}\left( \tau \right) - \frac{1}{10} \Pi\left( \tau \right) \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} G\_{3}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{7} \left( 3 G\_{2}\left( \tau \right) - 4 G\_{4}\left( \tau \right) \right) k + G\_{3}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} G\_{4}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{9} \left( 4 G\_{3}\left( \tau \right) - 5 G\_{5}\left( \tau \right) \right) k + G\_{4}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} G\_{5}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{1}{11} \left( 5 G\_{4}\left( \tau \right) - 6 G\_{6}\left( \tau \right) \right) k + G\_{5}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} G\_{6}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( G\_{5}\left( \tau \right) k + G\_{6}\left( \tau \right) \mathtt{\dot{\kappa}}\left( \tau \right) - 7 G\_{6}\left( \tau \right) \mathscr{E}\left( \tau \right) \right) \epsilon \\ T\left( \tau \right) &= \frac{\mathtt{T{_0}}}{a\left( \tau \right)} \\ w\left( \tau \right) &= \frac{1}{3} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \end{align*} \]

Initialization equations

initialization_equations(γ)

\[ \begin{align*} F\_{0}\left( \tau \right) \epsilon &= - 2 \Psi\left( \tau \right) \epsilon \\ F\_{1}\left( \tau \right) \epsilon &= \frac{\frac{2}{3} k \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right)} \\ F\_{2}\left( \tau \right) \epsilon &= \frac{ - \frac{8}{15} F\_{1}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ F\_{3}\left( \tau \right) \epsilon &= \frac{ - \frac{3}{7} F\_{2}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ F\_{4}\left( \tau \right) \epsilon &= \frac{ - \frac{4}{9} F\_{3}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ F\_{5}\left( \tau \right) \epsilon &= \frac{ - \frac{5}{11} F\_{4}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ F\_{6}\left( \tau \right) \epsilon &= \frac{ - \frac{6}{13} F\_{5}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{0}\left( \tau \right) \epsilon &= \frac{5}{16} F\_{2}\left( \tau \right) \epsilon \\ G\_{1}\left( \tau \right) \epsilon &= \frac{ - \frac{1}{16} F\_{2}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{2}\left( \tau \right) \epsilon &= \frac{1}{16} F\_{2}\left( \tau \right) \epsilon \\ G\_{3}\left( \tau \right) \epsilon &= \frac{ - \frac{3}{7} G\_{2}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{4}\left( \tau \right) \epsilon &= \frac{ - \frac{4}{9} G\_{3}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{5}\left( \tau \right) \epsilon &= \frac{ - \frac{5}{11} G\_{4}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \\ G\_{6}\left( \tau \right) \epsilon &= \frac{ - \frac{6}{13} G\_{5}\left( \tau \right) k \epsilon}{\mathtt{\dot{\kappa}}\left( \tau \right)} \end{align*} \]

Non-relativistic

SymBoltz.matter — Function

matter(g; name = :m, kwargs...)

Create a particle species for matter (with equation of state w ~ 0) in the spacetime with metric g.

m = SymBoltz.matter(g)

Equations

equations(m)

\[ \begin{align*} w\left( \tau \right) &= 0 \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ \frac{\mathrm{d} \delta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau \right) - 3 \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{k^{2} \mathtt{c_s^2}\left( \tau \right) \delta\left( \tau \right)}{1 + w\left( \tau \right)} + \mathtt{{\theta}interaction}\left( \tau \right) - k^{2} \sigma\left( \tau \right) + k^{2} \Psi\left( \tau \right) - \left( 1 - 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau \right) \right) \epsilon \\ u\left( \tau \right) \epsilon &= \frac{\theta\left( \tau \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( \tau \right) \epsilon &= \frac{\mathrm{d} u\left( \tau \right)}{\mathrm{d}\tau} \epsilon \\ \Delta\left( \tau \right) \epsilon &= \left( \frac{3 \left( 1 + w\left( \tau \right) \right) \mathscr{E}\left( \tau \right)}{\theta\left( \tau \right)} + \delta\left( \tau \right) \right) \epsilon \\ \sigma\left( \tau \right) \epsilon &= 0 \\ \mathtt{{\theta}interaction}\left( \tau \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(m)

\[ \begin{align*} \delta\left( \tau \right) \epsilon &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.cold_dark_matter — Function

cold_dark_matter(g; name = :c, kwargs...)

Create a particle species for cold dark matter in the spacetime with metric g.

c = SymBoltz.cold_dark_matter(g)

Equations

equations(c)

\[ \begin{align*} \mathtt{c_s^2}\left( \tau \right) \epsilon &= 0 \\ w\left( \tau \right) &= 0 \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ \frac{\mathrm{d} \delta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau \right) - 3 \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{k^{2} \mathtt{c_s^2}\left( \tau \right) \delta\left( \tau \right)}{1 + w\left( \tau \right)} + \mathtt{{\theta}interaction}\left( \tau \right) - k^{2} \sigma\left( \tau \right) + k^{2} \Psi\left( \tau \right) - \left( 1 - 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau \right) \right) \epsilon \\ u\left( \tau \right) \epsilon &= \frac{\theta\left( \tau \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( \tau \right) \epsilon &= \frac{\mathrm{d} u\left( \tau \right)}{\mathrm{d}\tau} \epsilon \\ \Delta\left( \tau \right) \epsilon &= \left( \frac{3 \left( 1 + w\left( \tau \right) \right) \mathscr{E}\left( \tau \right)}{\theta\left( \tau \right)} + \delta\left( \tau \right) \right) \epsilon \\ \sigma\left( \tau \right) \epsilon &= 0 \\ \mathtt{{\theta}interaction}\left( \tau \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(c)

\[ \begin{align*} \delta\left( \tau \right) \epsilon &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.baryons — Function

baryons(g; recombination=true, name = :b, kwargs...)

Create a particle species for baryons in the spacetime with metric g.

b = SymBoltz.baryons(g)

Equations

equations(b)

\[ \begin{align*} \mathtt{c_s^2}\left( \tau \right) \epsilon &= \mathtt{rec.c_s^2}\left( \tau \right) \epsilon \\ w\left( \tau \right) &= 0 \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ \frac{\mathrm{d} \delta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau \right) - 3 \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} \right) - 3 \left( - w\left( \tau \right) + \mathtt{c_s^2}\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{k^{2} \mathtt{c_s^2}\left( \tau \right) \delta\left( \tau \right)}{1 + w\left( \tau \right)} + \mathtt{{\theta}interaction}\left( \tau \right) - k^{2} \sigma\left( \tau \right) + k^{2} \Psi\left( \tau \right) - \left( 1 - 3 w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \theta\left( \tau \right) \right) \epsilon \\ u\left( \tau \right) \epsilon &= \frac{\theta\left( \tau \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( \tau \right) \epsilon &= \frac{\mathrm{d} u\left( \tau \right)}{\mathrm{d}\tau} \epsilon \\ \Delta\left( \tau \right) \epsilon &= \left( \frac{3 \left( 1 + w\left( \tau \right) \right) \mathscr{E}\left( \tau \right)}{\theta\left( \tau \right)} + \delta\left( \tau \right) \right) \epsilon \\ \sigma\left( \tau \right) \epsilon &= 0 \\ \mathtt{rec.nH}\left( \tau \right) &= 5.9743 \cdot 10^{26} \left( 1 - \mathtt{rec.Yp} \right) \mathtt{rec.{\rho}b}\left( \tau \right) \\ \mathtt{rec.nHe}\left( \tau \right) &= \mathtt{rec.fHe} \mathtt{rec.nH}\left( \tau \right) \\ \mathtt{rec.DTb}\left( \tau \right) &= \frac{ - 0.00015165 \left( \mathtt{rec.T\gamma}\left( \tau \right) \right)^{4} a\left( \tau \right) \mathtt{rec.{\Delta}T}\left( \tau \right) \mathtt{rec.Xe}\left( \tau \right)}{h \left( 1 + \mathtt{rec.fHe} + \mathtt{rec.Xe}\left( \tau \right) \right)} - 2 \mathtt{rec.Tb}\left( \tau \right) \mathscr{E}\left( \tau \right) \\ \mathtt{rec.DT\gamma}\left( \tau \right) &= \frac{\mathrm{d} \mathtt{rec.T\gamma}\left( \tau \right)}{\mathrm{d}\tau} \\ \frac{\mathrm{d} \mathtt{rec.{\Delta}T}\left( \tau \right)}{\mathrm{d}\tau} &= \mathtt{rec.DTb}\left( \tau \right) - \mathtt{rec.DT\gamma}\left( \tau \right) \\ \mathtt{rec.Tb}\left( \tau \right) &= \mathtt{rec.{\Delta}T}\left( \tau \right) + \mathtt{rec.T\gamma}\left( \tau \right) \\ \mathtt{rec.{\beta}b}\left( \tau \right) &= \frac{1}{1.3806 \cdot 10^{-23} \mathtt{rec.Tb}\left( \tau \right)} \\ \mathtt{rec.{\lambda}e}\left( \tau \right) &= \frac{6.6261 \cdot 10^{-34}}{\sqrt{\frac{5.7236 \cdot 10^{-30}}{\mathtt{rec.{\beta}b}\left( \tau \right)}}} \\ \mathtt{rec.\mu}\left( \tau \right) &= \frac{1.6738 \cdot 10^{-27}}{1 - 0.74816 \mathtt{rec.Yp} + \left( 1 - \mathtt{rec.Yp} \right) \mathtt{rec.Xe}\left( \tau \right)} \\ \mathtt{rec.c_s^2}\left( \tau \right) &= \frac{1.3806 \cdot 10^{-23} \left( \mathtt{rec.Tb}\left( \tau \right) + \frac{ - \frac{\mathrm{d} \mathtt{rec.Tb}\left( \tau \right)}{\mathrm{d}\tau}}{3 \mathscr{E}\left( \tau \right)} \right)}{8.9876 \cdot 10^{16} \mathtt{rec.\mu}\left( \tau \right)} \\ \mathtt{rec.{\alpha}H}\left( \tau \right) &= \frac{4.9123 \cdot 10^{-19}}{0.0034166 \left( \mathtt{rec.Tb}\left( \tau \right) \right)^{0.6166} \left( 1 + 0.0050847 \left( \mathtt{rec.Tb}\left( \tau \right) \right)^{0.53} \right)} \\ \mathtt{rec.{\beta}H}\left( \tau \right) &= \frac{\mathtt{rec.{\alpha}H}\left( \tau \right) e^{ - 5.4468 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)}}{\left( \mathtt{rec.{\lambda}e}\left( \tau \right) \right)^{3}} \\ \mathtt{rec.KH0}\left( \tau \right) &= \frac{1.7966 \cdot 10^{-21}}{25.133 H\left( \tau \right)} \\ \mathtt{rec.KH1}\left( \tau \right) &= \left( 0.079 e^{ - 9.1827 \left( 6.73 + \log\left( a\left( \tau \right) \right) \right)^{2}} - 0.14 e^{ - 30.864 \left( 7.28 + \log\left( a\left( \tau \right) \right) \right)^{2}} \right) \mathtt{rec.KH0}\left( \tau \right) \\ \mathtt{rec.KH}\left( \tau \right) &= \mathtt{rec.KH1}\left( \tau \right) + \mathtt{rec.KH0}\left( \tau \right) \\ \mathtt{rec.CH}\left( \tau \right) &= \frac{1 + 8.2246 \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right) \mathtt{rec.KH}\left( \tau \right) \mathtt{rec.nH}\left( \tau \right)}{1 + \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right) \mathtt{rec.KH}\left( \tau \right) \left( 8.2246 + \mathtt{rec.{\beta}H}\left( \tau \right) \right) \mathtt{rec.nH}\left( \tau \right)} \\ \frac{\mathrm{d} \mathtt{rec.XH^+}\left( \tau \right)}{\mathrm{d}\tau} &= \frac{ - a\left( \tau \right) \left( \mathtt{rec.XH^+}\left( \tau \right) \mathtt{rec.{\alpha}H}\left( \tau \right) \mathtt{rec.ne}\left( \tau \right) - \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right) \mathtt{rec.{\beta}H}\left( \tau \right) e^{ - 1.634 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( \tau \right)} \right) \mathtt{rec.CH}\left( \tau \right)}{3.2408 \cdot 10^{-18} h} \\ \mathtt{rec.{\alpha}He}\left( \tau \right) &= \frac{1.803 \cdot 10^{-17}}{\left( 1 + \sqrt{7.6913 \cdot 10^{-6} \mathtt{rec.Tb}\left( \tau \right)} \right)^{1.711} \left( 1 + \sqrt{0.33333 \mathtt{rec.Tb}\left( \tau \right)} \right)^{0.289} \sqrt{0.33333 \mathtt{rec.Tb}\left( \tau \right)}} \\ \mathtt{rec.{\beta}He}\left( \tau \right) &= \frac{4 e^{ - 6.3633 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)} \mathtt{rec.{\alpha}He}\left( \tau \right)}{\left( \mathtt{rec.{\lambda}e}\left( \tau \right) \right)^{3}} \\ \mathtt{rec.KHe0^{- 1}}\left( \tau \right) &= 1.2597 \cdot 10^{23} H\left( \tau \right) \\ \mathtt{rec.KHe1^{- 1}}\left( \tau \right) &= - e^{\frac{ - 5.3949 \cdot 10^{9} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right)}{\mathtt{rec.KHe0^{- 1}}\left( \tau \right)}} \mathtt{rec.KHe0^{- 1}}\left( \tau \right) \\ \mathtt{rec.{\gamma}2Ps}\left( \tau \right) &= \left|\frac{4.8487 \cdot 10^{26} \mathtt{rec.fHe} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right)}{4.8748 \cdot 10^{26} \sqrt{\frac{6.2832}{5.9736 \cdot 10^{-10} \mathtt{rec.{\beta}b}\left( \tau \right)}} \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right)}\right| \\ \mathtt{rec.KHe2^{- 1}}\left( \tau \right) &= \frac{5.3949 \cdot 10^{9} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right)}{1 + 0.36 \left( \mathtt{rec.{\gamma}2Ps}\left( \tau \right) \right)^{0.86}} \\ \mathtt{rec.KHe}\left( \tau \right) &= \frac{1}{\mathtt{rec.KHe2^{- 1}}\left( \tau \right) + \mathtt{rec.KHe0^{- 1}}\left( \tau \right) + \mathtt{rec.KHe1^{- 1}}\left( \tau \right)} \\ \mathtt{rec.CHe}\left( \tau \right) &= \frac{e^{ - 9.6491 \cdot 10^{-20} \mathtt{rec.{\beta}b}\left( \tau \right)} + 51.3 \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right) \mathtt{rec.KHe}\left( \tau \right)}{e^{ - 9.6491 \cdot 10^{-20} \mathtt{rec.{\beta}b}\left( \tau \right)} + \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \left( 51.3 + \mathtt{rec.{\beta}He}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right) \mathtt{rec.KHe}\left( \tau \right)} \\ \mathtt{rec.DXHe^+\_singlet}\left( \tau \right) &= \frac{ - a\left( \tau \right) \left( - \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.{\beta}He}\left( \tau \right) e^{ - 3.303 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( \tau \right)} + \mathtt{rec.XHe^+}\left( \tau \right) \mathtt{rec.{\alpha}He}\left( \tau \right) \mathtt{rec.ne}\left( \tau \right) \right) \mathtt{rec.CHe}\left( \tau \right)}{3.2408 \cdot 10^{-18} h} \\ \mathtt{rec.{\alpha}He3}\left( \tau \right) &= \frac{4.9431 \cdot 10^{-17}}{\left( 1 + \sqrt{7.6913 \cdot 10^{-6} \mathtt{rec.Tb}\left( \tau \right)} \right)^{1.761} \left( 1 + \sqrt{0.33333 \mathtt{rec.Tb}\left( \tau \right)} \right)^{0.239} \sqrt{0.33333 \mathtt{rec.Tb}\left( \tau \right)}} \\ \mathtt{rec.{\beta}He3}\left( \tau \right) &= \frac{1.3333 e^{ - 7.6388 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)} \mathtt{rec.{\alpha}He3}\left( \tau \right)}{\left( \mathtt{rec.{\lambda}e}\left( \tau \right) \right)^{3}} \\ \mathtt{rec.{\tau}He3}\left( \tau \right) &= \frac{1.102 \cdot 10^{-19} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) \mathtt{rec.nHe}\left( \tau \right)}{25.133 H\left( \tau \right)} \\ \mathtt{rec.pHe3}\left( \tau \right) &= \frac{1 - e^{ - \mathtt{rec.{\tau}He3}\left( \tau \right)}}{\mathtt{rec.{\tau}He3}\left( \tau \right)} \\ \mathtt{rec.{\gamma}2Pt}\left( \tau \right) &= \left|\frac{1.8633 \cdot 10^{-12} \mathtt{rec.fHe} \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right)}{1.8917 \cdot 10^{-5} \sqrt{\frac{6.2832}{5.9736 \cdot 10^{-10} \mathtt{rec.{\beta}b}\left( \tau \right)}} \left( 1 - \mathtt{rec.XH^+}\left( \tau \right) \right)}\right| \\ \mathtt{rec.CHe3}\left( \tau \right) &= \frac{1 \cdot 10^{-10} + 177.58 \left( \mathtt{rec.pHe3}\left( \tau \right) + \frac{\frac{1}{3}}{1 + 0.66 \left( \mathtt{rec.{\gamma}2Pt}\left( \tau \right) \right)^{0.9}} \right) e^{ - 1.8337 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)}}{1 \cdot 10^{-10} + \mathtt{rec.{\beta}He3}\left( \tau \right) + 177.58 \left( \mathtt{rec.pHe3}\left( \tau \right) + \frac{\frac{1}{3}}{1 + 0.66 \left( \mathtt{rec.{\gamma}2Pt}\left( \tau \right) \right)^{0.9}} \right) e^{ - 1.8337 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( \tau \right)}} \\ \mathtt{rec.DXHe^+\_triplet}\left( \tau \right) &= \frac{ - a\left( \tau \right) \left( - 3 \left( 1 - \mathtt{rec.XHe^+}\left( \tau \right) \right) e^{ - 3.1755 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( \tau \right)} \mathtt{rec.{\beta}He3}\left( \tau \right) + \mathtt{rec.XHe^+}\left( \tau \right) \mathtt{rec.ne}\left( \tau \right) \mathtt{rec.{\alpha}He3}\left( \tau \right) \right) \mathtt{rec.CHe3}\left( \tau \right)}{3.2408 \cdot 10^{-18} h} \\ \frac{\mathrm{d} \mathtt{rec.XHe^+}\left( \tau \right)}{\mathrm{d}\tau} &= \mathtt{rec.DXHe^+\_triplet}\left( \tau \right) + \mathtt{rec.DXHe^+\_singlet}\left( \tau \right) \\ \mathtt{rec.RHe^+}\left( \tau \right) &= \frac{e^{ - 8.7187 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( \tau \right)}}{\left( \mathtt{rec.{\lambda}e}\left( \tau \right) \right)^{3} \mathtt{rec.nH}\left( \tau \right)} \\ \mathtt{rec.XHe^{+ +}}\left( \tau \right) &= \frac{2 \mathtt{rec.fHe} \mathtt{rec.RHe^+}\left( \tau \right)}{\left( 1 + \mathtt{rec.fHe} + \mathtt{rec.RHe^+}\left( \tau \right) \right) \left( 1 + \sqrt{1 + \frac{4 \mathtt{rec.fHe} \mathtt{rec.RHe^+}\left( \tau \right)}{\left( 1 + \mathtt{rec.fHe} + \mathtt{rec.RHe^+}\left( \tau \right) \right)^{2}}} \right)} \\ \mathtt{rec.re1Xe}\left( \tau \right) &= 0.5 \left( 1 + \mathtt{rec.fHe} + \left( 1 + \mathtt{rec.fHe} \right) \tanh\left( \frac{\left( 1 + \mathtt{rec.re1z} \right)^{1.5} - \left( 1 + z\left( \tau \right) \right)^{1.5}}{0.75 \left( 1 + \mathtt{rec.re1z} \right)^{0.5}} \right) \right) \\ \mathtt{rec.re2Xe}\left( \tau \right) &= 0.5 \left( \mathtt{rec.fHe} + \mathtt{rec.fHe} \tanh\left( \frac{\left( 1 + \mathtt{rec.re2z} \right)^{1.5} - \left( 1 + z\left( \tau \right) \right)^{1.5}}{0.75 \left( 1 + \mathtt{rec.re2z} \right)^{0.5}} \right) \right) \\ \mathtt{rec.Xe}\left( \tau \right) &= \mathtt{rec.XH^+}\left( \tau \right) + \mathtt{rec.re1Xe}\left( \tau \right) + \mathtt{rec.re2Xe}\left( \tau \right) + \mathtt{rec.XHe^{+ +}}\left( \tau \right) + \mathtt{rec.fHe} \mathtt{rec.XHe^+}\left( \tau \right) \\ \mathtt{rec.ne}\left( \tau \right) &= \mathtt{rec.Xe}\left( \tau \right) \mathtt{rec.nH}\left( \tau \right) \\ \frac{\mathrm{d} \mathtt{rec.\_\kappa}\left( \tau \right)}{\mathrm{d}\tau} &= \frac{ - 1.9944 \cdot 10^{-20} a\left( \tau \right) \mathtt{rec.ne}\left( \tau \right)}{3.2408 \cdot 10^{-18} h} \\ \mathtt{rec.\dot{\kappa}}\left( \tau \right) &= \frac{\mathrm{d} \mathtt{rec.\_\kappa}\left( \tau \right)}{\mathrm{d}\tau} \\ \mathtt{rec.\kappa}\left( \tau \right) &= - \mathtt{rec.{\kappa}0} + \mathtt{rec.\_\kappa}\left( \tau \right) \\ \mathtt{rec.v}\left( \tau \right) &= - \frac{\mathrm{d} \mathtt{rec.\kappa}\left( \tau \right)}{\mathrm{d}\tau} e^{ - \mathtt{rec.\kappa}\left( \tau \right)} \\ \mathtt{rec.\dot{v}}\left( \tau \right) &= \frac{\mathrm{d} \mathtt{rec.v}\left( \tau \right)}{\mathrm{d}\tau} \end{align*} \]

Initialization equations

initialization_equations(b)

\[ \begin{align*} \delta\left( \tau \right) \epsilon &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right)} \end{align*} \]

Neutrinos

SymBoltz.massless_neutrinos — Function

massless_neutrinos(g; lmax = 6, name = :ν, kwargs...)

Create a particle species for massless neutrinos in the spacetime with metric g.

ν = SymBoltz.massless_neutrinos(g; lmax = 6)

Equations

equations(ν)

\[ \begin{align*} \frac{\mathrm{d} F\_{0}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( 4 \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} - F\_{1}\left( \tau \right) k \right) \epsilon \\ \frac{\mathrm{d} F\_{1}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{1}{3} \left( F\_{0}\left( \tau \right) - 2 F\_{2}\left( \tau \right) + 4 \Psi\left( \tau \right) \right) k \epsilon \\ \frac{\mathrm{d} F\_{2}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{1}{5} \left( 2 F\_{1}\left( \tau \right) - 3 F\_{3}\left( \tau \right) \right) k \epsilon \\ \frac{\mathrm{d} F\_{3}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{1}{7} \left( 3 F\_{2}\left( \tau \right) - 4 F\_{4}\left( \tau \right) \right) k \epsilon \\ \frac{\mathrm{d} F\_{4}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{1}{9} \left( 4 F\_{3}\left( \tau \right) - 5 F\_{5}\left( \tau \right) \right) k \epsilon \\ \frac{\mathrm{d} F\_{5}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{1}{11} \left( 5 F\_{4}\left( \tau \right) - 6 F\_{6}\left( \tau \right) \right) k \epsilon \\ \frac{\mathrm{d} F\_{6}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( F\_{5}\left( \tau \right) k - 7 F\_{6}\left( \tau \right) \mathscr{E}\left( \tau \right) \right) \epsilon \\ \delta\left( \tau \right) \epsilon &= F\_{0}\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{3}{4} F\_{1}\left( \tau \right) k \epsilon \\ \sigma\left( \tau \right) \epsilon &= \frac{1}{2} F\_{2}\left( \tau \right) \epsilon \\ \mathtt{c_s^2}\left( \tau \right) \epsilon &= \frac{1}{3} \epsilon \\ T\left( \tau \right) &= \frac{\mathtt{T{_0}}}{a\left( \tau \right)} \\ w\left( \tau \right) &= \frac{1}{3} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \end{align*} \]

Initialization equations

initialization_equations(ν)

\[ \begin{align*} \delta\left( \tau \right) \epsilon &= - 2 \Psi\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right)} \\ \sigma\left( \tau \right) \epsilon &= \frac{1}{15} \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} \Psi\left( \tau \right) \epsilon \\ F\_{3}\left( \tau \right) \epsilon &= \frac{\frac{3}{7} F\_{2}\left( \tau \right) k \epsilon}{\mathscr{E}\left( \tau \right)} \\ F\_{4}\left( \tau \right) \epsilon &= \frac{\frac{4}{9} F\_{3}\left( \tau \right) k \epsilon}{\mathscr{E}\left( \tau \right)} \\ F\_{5}\left( \tau \right) \epsilon &= \frac{\frac{5}{11} F\_{4}\left( \tau \right) k \epsilon}{\mathscr{E}\left( \tau \right)} \\ F\_{6}\left( \tau \right) \epsilon &= \frac{\frac{6}{13} F\_{5}\left( \tau \right) k \epsilon}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.massive_neutrinos — Function

massive_neutrinos(g; nx = 5, lmax = 4, name = :h, kwargs...)

Create a particle species for massive neutrinos in the spacetime with metric g.

h = SymBoltz.massive_neutrinos(g; nx=4, lmax = 5)

Equations

equations(h)

\[ \begin{align*} T\left( \tau \right) &= \frac{\mathtt{T{_0}}}{a\left( \tau \right)} \\ y\left( \tau \right) &= \frac{8.9876 \cdot 10^{16} m}{1.3806 \cdot 10^{-23} T\left( \tau \right)} \\ \mathtt{In}\left( \tau \right) &= W_{1} + W_{2} + W_{3} + W_{4} \\ \mathtt{I\rho}\left( \tau \right) &= E\_{1}\left( \tau \right) W_{1} + E\_{2}\left( \tau \right) W_{2} + E\_{3}\left( \tau \right) W_{3} + E\_{4}\left( \tau \right) W_{4} \\ \mathtt{IP}\left( \tau \right) &= \frac{\left( x_{3} \right)^{2} W_{3}}{E\_{3}\left( \tau \right)} + \frac{\left( x_{4} \right)^{2} W_{4}}{E\_{4}\left( \tau \right)} + \frac{\left( x_{2} \right)^{2} W_{2}}{E\_{2}\left( \tau \right)} + \frac{\left( x_{1} \right)^{2} W_{1}}{E\_{1}\left( \tau \right)} \\ \rho\left( \tau \right) &= \frac{1.165 \cdot 10^{-16} \left( T\left( \tau \right) \right)^{4} \mathtt{I\rho}\left( \tau \right)}{1.4143 \cdot 10^{-8} h^{2}} \\ P\left( \tau \right) &= \frac{3.8835 \cdot 10^{-17} \left( T\left( \tau \right) \right)^{4} \mathtt{IP}\left( \tau \right)}{1.4143 \cdot 10^{-8} h^{2}} \\ w\left( \tau \right) &= \frac{P\left( \tau \right)}{\rho\left( \tau \right)} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \\ E\_{1}\left( \tau \right) &= \sqrt{\left( x_{1} \right)^{2} + \left( y\left( \tau \right) \right)^{2}} \\ E\_{2}\left( \tau \right) &= \sqrt{\left( x_{2} \right)^{2} + \left( y\left( \tau \right) \right)^{2}} \\ E\_{3}\left( \tau \right) &= \sqrt{\left( x_{3} \right)^{2} + \left( y\left( \tau \right) \right)^{2}} \\ E\_{4}\left( \tau \right) &= \sqrt{\left( x_{4} \right)^{2} + \left( y\left( \tau \right) \right)^{2}} \\ \mathtt{I\delta\rho}\left( \tau \right) \epsilon &= \left( E\_{1}\left( \tau \right) W_{1} \psi\_{1,0}\left( \tau \right) + E\_{2}\left( \tau \right) W_{2} \psi\_{2,0}\left( \tau \right) + E\_{3}\left( \tau \right) W_{3} \psi\_{3,0}\left( \tau \right) + E\_{4}\left( \tau \right) W_{4} \psi\_{4,0}\left( \tau \right) \right) \epsilon \\ \delta\left( \tau \right) \epsilon &= \frac{\mathtt{I\delta\rho}\left( \tau \right) \epsilon}{\mathtt{I\rho}\left( \tau \right)} \\ u\left( \tau \right) \epsilon &= \frac{\left( W_{1} x_{1} \psi\_{1,1}\left( \tau \right) + W_{2} x_{2} \psi\_{2,1}\left( \tau \right) + W_{3} x_{3} \psi\_{3,1}\left( \tau \right) + W_{4} x_{4} \psi\_{4,1}\left( \tau \right) \right) \epsilon}{\mathtt{I\rho}\left( \tau \right) + \frac{1}{3} \mathtt{IP}\left( \tau \right)} \\ \theta\left( \tau \right) \epsilon &= k u\left( \tau \right) \epsilon \\ \sigma\left( \tau \right) \epsilon &= \frac{\frac{2}{3} \left( \frac{\left( x_{3} \right)^{2} W_{3} \psi\_{3,2}\left( \tau \right)}{E\_{3}\left( \tau \right)} + \frac{\left( x_{4} \right)^{2} W_{4} \psi\_{4,2}\left( \tau \right)}{E\_{4}\left( \tau \right)} + \frac{\left( x_{2} \right)^{2} W_{2} \psi\_{2,2}\left( \tau \right)}{E\_{2}\left( \tau \right)} + \frac{\left( x_{1} \right)^{2} W_{1} \psi\_{1,2}\left( \tau \right)}{E\_{1}\left( \tau \right)} \right) \epsilon}{\mathtt{I\rho}\left( \tau \right) + \frac{1}{3} \mathtt{IP}\left( \tau \right)} \\ \mathtt{c_s^2}\left( \tau \right) \epsilon &= \frac{\left( \frac{\left( x_{3} \right)^{2} W_{3} \psi\_{3,0}\left( \tau \right)}{E\_{3}\left( \tau \right)} + \frac{\left( x_{2} \right)^{2} W_{2} \psi\_{2,0}\left( \tau \right)}{E\_{2}\left( \tau \right)} + \frac{\left( x_{4} \right)^{2} W_{4} \psi\_{4,0}\left( \tau \right)}{E\_{4}\left( \tau \right)} + \frac{\left( x_{1} \right)^{2} W_{1} \psi\_{1,0}\left( \tau \right)}{E\_{1}\left( \tau \right)} \right) \epsilon}{\mathtt{I\delta\rho}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{1,0}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{ - k x_{1} \psi\_{1,1}\left( \tau \right)}{E\_{1}\left( \tau \right)} + \frac{x_{1} \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau}}{1 + e^{ - x_{1}}} \right) \epsilon \\ \frac{\mathrm{d} \psi\_{1,1}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{\frac{1}{3} k x_{1} \left( \psi\_{1,0}\left( \tau \right) - 2 \psi\_{1,2}\left( \tau \right) \right)}{E\_{1}\left( \tau \right)} + \frac{\frac{1}{3} E\_{1}\left( \tau \right) k \Psi\left( \tau \right)}{1 + e^{ - x_{1}}} \right) \epsilon \\ \frac{\mathrm{d} \psi\_{1,2}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{5} k x_{1} \left( 2 \psi\_{1,1}\left( \tau \right) - 3 \psi\_{1,3}\left( \tau \right) \right) \epsilon}{E\_{1}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{1,3}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{7} k x_{1} \left( 3 \psi\_{1,2}\left( \tau \right) - 4 \psi\_{1,4}\left( \tau \right) \right) \epsilon}{E\_{1}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{1,4}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{9} k x_{1} \left( 4 \psi\_{1,3}\left( \tau \right) - 5 \psi\_{1,5}\left( \tau \right) \right) \epsilon}{E\_{1}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{1,5}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{11} k x_{1} \left( 5 \psi\_{1,4}\left( \tau \right) - 6 \psi\_{1,6}\left( \tau \right) \right) \epsilon}{E\_{1}\left( \tau \right)} \\ \psi\_{1,6}\left( \tau \right) \epsilon &= \left( \frac{11 E\_{1}\left( \tau \right) \mathscr{E}\left( \tau \right) \psi\_{1,5}\left( \tau \right)}{k x_{1}} - \psi\_{1,4}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \psi\_{2,0}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{ - k x_{2} \psi\_{2,1}\left( \tau \right)}{E\_{2}\left( \tau \right)} + \frac{x_{2} \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau}}{1 + e^{ - x_{2}}} \right) \epsilon \\ \frac{\mathrm{d} \psi\_{2,1}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{\frac{1}{3} E\_{2}\left( \tau \right) k \Psi\left( \tau \right)}{1 + e^{ - x_{2}}} + \frac{\frac{1}{3} k x_{2} \left( \psi\_{2,0}\left( \tau \right) - 2 \psi\_{2,2}\left( \tau \right) \right)}{E\_{2}\left( \tau \right)} \right) \epsilon \\ \frac{\mathrm{d} \psi\_{2,2}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{5} k x_{2} \left( 2 \psi\_{2,1}\left( \tau \right) - 3 \psi\_{2,3}\left( \tau \right) \right) \epsilon}{E\_{2}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{2,3}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{7} k x_{2} \left( 3 \psi\_{2,2}\left( \tau \right) - 4 \psi\_{2,4}\left( \tau \right) \right) \epsilon}{E\_{2}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{2,4}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{9} k x_{2} \left( 4 \psi\_{2,3}\left( \tau \right) - 5 \psi\_{2,5}\left( \tau \right) \right) \epsilon}{E\_{2}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{2,5}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{11} k x_{2} \left( 5 \psi\_{2,4}\left( \tau \right) - 6 \psi\_{2,6}\left( \tau \right) \right) \epsilon}{E\_{2}\left( \tau \right)} \\ \psi\_{2,6}\left( \tau \right) \epsilon &= \left( \frac{11 E\_{2}\left( \tau \right) \mathscr{E}\left( \tau \right) \psi\_{2,5}\left( \tau \right)}{k x_{2}} - \psi\_{2,4}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \psi\_{3,0}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{ - k x_{3} \psi\_{3,1}\left( \tau \right)}{E\_{3}\left( \tau \right)} + \frac{x_{3} \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau}}{1 + e^{ - x_{3}}} \right) \epsilon \\ \frac{\mathrm{d} \psi\_{3,1}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{\frac{1}{3} k x_{3} \left( \psi\_{3,0}\left( \tau \right) - 2 \psi\_{3,2}\left( \tau \right) \right)}{E\_{3}\left( \tau \right)} + \frac{\frac{1}{3} E\_{3}\left( \tau \right) k \Psi\left( \tau \right)}{1 + e^{ - x_{3}}} \right) \epsilon \\ \frac{\mathrm{d} \psi\_{3,2}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{5} k x_{3} \left( 2 \psi\_{3,1}\left( \tau \right) - 3 \psi\_{3,3}\left( \tau \right) \right) \epsilon}{E\_{3}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{3,3}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{7} k x_{3} \left( 3 \psi\_{3,2}\left( \tau \right) - 4 \psi\_{3,4}\left( \tau \right) \right) \epsilon}{E\_{3}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{3,4}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{9} k x_{3} \left( 4 \psi\_{3,3}\left( \tau \right) - 5 \psi\_{3,5}\left( \tau \right) \right) \epsilon}{E\_{3}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{3,5}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{11} k x_{3} \left( 5 \psi\_{3,4}\left( \tau \right) - 6 \psi\_{3,6}\left( \tau \right) \right) \epsilon}{E\_{3}\left( \tau \right)} \\ \psi\_{3,6}\left( \tau \right) \epsilon &= \left( \frac{11 E\_{3}\left( \tau \right) \mathscr{E}\left( \tau \right) \psi\_{3,5}\left( \tau \right)}{k x_{3}} - \psi\_{3,4}\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \psi\_{4,0}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{x_{4} \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau}}{1 + e^{ - x_{4}}} + \frac{ - k x_{4} \psi\_{4,1}\left( \tau \right)}{E\_{4}\left( \tau \right)} \right) \epsilon \\ \frac{\mathrm{d} \psi\_{4,1}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{\frac{1}{3} E\_{4}\left( \tau \right) k \Psi\left( \tau \right)}{1 + e^{ - x_{4}}} + \frac{\frac{1}{3} k x_{4} \left( \psi\_{4,0}\left( \tau \right) - 2 \psi\_{4,2}\left( \tau \right) \right)}{E\_{4}\left( \tau \right)} \right) \epsilon \\ \frac{\mathrm{d} \psi\_{4,2}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{5} k x_{4} \left( 2 \psi\_{4,1}\left( \tau \right) - 3 \psi\_{4,3}\left( \tau \right) \right) \epsilon}{E\_{4}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{4,3}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{7} k x_{4} \left( 3 \psi\_{4,2}\left( \tau \right) - 4 \psi\_{4,4}\left( \tau \right) \right) \epsilon}{E\_{4}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{4,4}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{9} k x_{4} \left( 4 \psi\_{4,3}\left( \tau \right) - 5 \psi\_{4,5}\left( \tau \right) \right) \epsilon}{E\_{4}\left( \tau \right)} \\ \frac{\mathrm{d} \psi\_{4,5}\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \frac{\frac{1}{11} k x_{4} \left( 5 \psi\_{4,4}\left( \tau \right) - 6 \psi\_{4,6}\left( \tau \right) \right) \epsilon}{E\_{4}\left( \tau \right)} \\ \psi\_{4,6}\left( \tau \right) \epsilon &= \left( \frac{11 E\_{4}\left( \tau \right) \mathscr{E}\left( \tau \right) \psi\_{4,5}\left( \tau \right)}{k x_{4}} - \psi\_{4,4}\left( \tau \right) \right) \epsilon \end{align*} \]

Initialization equations

initialization_equations(h)

\[ \begin{align*} \psi\_{1,0}\left( \tau \right) \epsilon &= \frac{ - \frac{1}{2} x_{1} \Psi\left( \tau \right) \epsilon}{1 + e^{ - x_{1}}} \\ \psi\_{1,1}\left( \tau \right) \epsilon &= \frac{0.16667 E\_{1}\left( \tau \right) k \Psi\left( \tau \right) \epsilon}{\left( 1 + e^{ - x_{1}} \right) \mathscr{E}\left( \tau \right)} \\ \psi\_{1,2}\left( \tau \right) \epsilon &= \frac{\frac{1}{30} \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} x_{1} \Psi\left( \tau \right) \epsilon}{1 + e^{ - x_{1}}} \\ \psi\_{1,3}\left( \tau \right) \epsilon &= 0 \\ \psi\_{1,4}\left( \tau \right) \epsilon &= 0 \\ \psi\_{1,5}\left( \tau \right) \epsilon &= 0 \\ \psi\_{2,0}\left( \tau \right) \epsilon &= \frac{ - \frac{1}{2} x_{2} \Psi\left( \tau \right) \epsilon}{1 + e^{ - x_{2}}} \\ \psi\_{2,1}\left( \tau \right) \epsilon &= \frac{0.16667 E\_{2}\left( \tau \right) k \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right) \left( 1 + e^{ - x_{2}} \right)} \\ \psi\_{2,2}\left( \tau \right) \epsilon &= \frac{\frac{1}{30} \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} x_{2} \Psi\left( \tau \right) \epsilon}{1 + e^{ - x_{2}}} \\ \psi\_{2,3}\left( \tau \right) \epsilon &= 0 \\ \psi\_{2,4}\left( \tau \right) \epsilon &= 0 \\ \psi\_{2,5}\left( \tau \right) \epsilon &= 0 \\ \psi\_{3,0}\left( \tau \right) \epsilon &= \frac{ - \frac{1}{2} x_{3} \Psi\left( \tau \right) \epsilon}{1 + e^{ - x_{3}}} \\ \psi\_{3,1}\left( \tau \right) \epsilon &= \frac{0.16667 E\_{3}\left( \tau \right) k \Psi\left( \tau \right) \epsilon}{\left( 1 + e^{ - x_{3}} \right) \mathscr{E}\left( \tau \right)} \\ \psi\_{3,2}\left( \tau \right) \epsilon &= \frac{\frac{1}{30} \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} x_{3} \Psi\left( \tau \right) \epsilon}{1 + e^{ - x_{3}}} \\ \psi\_{3,3}\left( \tau \right) \epsilon &= 0 \\ \psi\_{3,4}\left( \tau \right) \epsilon &= 0 \\ \psi\_{3,5}\left( \tau \right) \epsilon &= 0 \\ \psi\_{4,0}\left( \tau \right) \epsilon &= \frac{ - \frac{1}{2} x_{4} \Psi\left( \tau \right) \epsilon}{1 + e^{ - x_{4}}} \\ \psi\_{4,1}\left( \tau \right) \epsilon &= \frac{0.16667 E\_{4}\left( \tau \right) k \Psi\left( \tau \right) \epsilon}{\left( 1 + e^{ - x_{4}} \right) \mathscr{E}\left( \tau \right)} \\ \psi\_{4,2}\left( \tau \right) \epsilon &= \frac{\frac{1}{30} \left( \frac{k}{\mathscr{E}\left( \tau \right)} \right)^{2} x_{4} \Psi\left( \tau \right) \epsilon}{1 + e^{ - x_{4}}} \\ \psi\_{4,3}\left( \tau \right) \epsilon &= 0 \\ \psi\_{4,4}\left( \tau \right) \epsilon &= 0 \\ \psi\_{4,5}\left( \tau \right) \epsilon &= 0 \end{align*} \]

Dark energy

SymBoltz.cosmological_constant — Function

cosmological_constant(g; name = :Λ, kwargs...)

Create a particle species for the cosmological constant (with equation of state w ~ -1) in the spacetime with metric g.

Λ = SymBoltz.cosmological_constant(g)

Equations

equations(Λ)

\[ \begin{align*} \delta\left( \tau \right) \epsilon &= 0 \\ \theta\left( \tau \right) \epsilon &= 0 \\ \sigma\left( \tau \right) \epsilon &= 0 \\ \mathtt{c_s^2}\left( \tau \right) \epsilon &= - \epsilon \\ w\left( \tau \right) &= -1 \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \rho\left( \tau \right) &= \frac{3 \left( a\left( \tau \right) \right)^{ - 3 \left( 1 + w\left( \tau \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( \tau \right) &= \frac{8}{3} \rho\left( \tau \right) \pi \end{align*} \]

Initialization equations

initialization_equations(Λ)

\[ \begin{align*} \end{align*} \]

SymBoltz.w0wa — Function

w0wa(g; kwargs...)

Create a particle species for the w₀-wₐ dark energy (CPL) parametrization in the spacetime with metric g.

X = SymBoltz.w0wa(g)

Equations

equations(X)

\[ \begin{align*} w\left( \tau \right) &= \mathtt{w0} + \mathtt{wa} \left( 1 - a\left( \tau \right) \right) \\ \dot{w}\left( \tau \right) &= \frac{\mathrm{d} w\left( \tau \right)}{\mathrm{d}\tau} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \frac{\mathrm{d} \rho\left( \tau \right)}{\mathrm{d}\tau} &= - 3 \left( 1 + w\left( \tau \right) \right) \rho\left( \tau \right) \mathscr{E}\left( \tau \right) \\ \mathtt{c_a^2}\left( \tau \right) \epsilon &= \left( w\left( \tau \right) + \frac{ - \dot{w}\left( \tau \right)}{3 \left( 1 + w\left( \tau \right) \right) \mathscr{E}\left( \tau \right)} \right) \epsilon \\ \frac{\mathrm{d} \delta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \left( -1 - w\left( \tau \right) \right) \left( \theta\left( \tau \right) - 3 \frac{\mathrm{d} \Phi\left( \tau \right)}{\mathrm{d}\tau} \right) - 3 \left( \mathtt{c_s^2} - w\left( \tau \right) \right) \mathscr{E}\left( \tau \right) \delta\left( \tau \right) - 9 \left( \frac{\mathscr{E}\left( \tau \right)}{k} \right)^{2} \left( \mathtt{c_s^2} - \mathtt{c_a^2}\left( \tau \right) \right) \left( 1 + w\left( \tau \right) \right) \theta\left( \tau \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( \tau \right)}{\mathrm{d}\tau} \epsilon &= \left( \frac{k^{2} \mathtt{c_s^2} \delta\left( \tau \right)}{1 + w\left( \tau \right)} - \left( 1 - 3 \mathtt{c_s^2} \right) \mathscr{E}\left( \tau \right) \theta\left( \tau \right) - k^{2} \sigma\left( \tau \right) + k^{2} \Psi\left( \tau \right) \right) \epsilon \\ \sigma\left( \tau \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(X)

\[ \begin{align*} \delta\left( \tau \right) \epsilon &= - \frac{3}{2} \left( 1 + w\left( \tau \right) \right) \Psi\left( \tau \right) \epsilon \\ \theta\left( \tau \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( \tau \right) \epsilon}{\mathscr{E}\left( \tau \right)} \end{align*} \]

SymBoltz.quintessence — Function

quintessence(g, v; name = :ϕ, kwargs...)

Create a species for a quintessence scalar field with potential v in the spacetime with metric g.

@variables ϕ(τ) V(ϕ) V′(ϕ) V′′(ϕ)
Q = SymBoltz.quintessence(g, V, V′, V′′)

Equations

equations(Q)

\[ \begin{align*} \frac{\mathrm{d}}{\mathrm{d}\tau} \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} &= - 2 \frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau} \mathscr{E}\left( \tau \right) - \left( a\left( \tau \right) \right)^{2} \mathtt{V\prime}\left( \tau \right) \\ V\left( \tau \right) &= V\left( \phi\left( \tau \right) \right) \\ \mathtt{V\prime}\left( \tau \right) &= \mathtt{V\prime}\left( \phi\left( \tau \right) \right) \\ \mathtt{V\prime\prime}\left( \tau \right) &= \mathtt{V\prime\prime}\left( \phi\left( \tau \right) \right) \\ K\left( \tau \right) &= \frac{1}{2} \left( \frac{\frac{\mathrm{d} \phi\left( \tau \right)}{\mathrm{d}\tau}}{a\left( \tau \right)} \right)^{2} \\ \rho\left( \tau \right) &= V\left( \tau \right) + K\left( \tau \right) \\ P\left( \tau \right) &= - V\left( \tau \right) + K\left( \tau \right) \\ w\left( \tau \right) &= \frac{P\left( \tau \right)}{\rho\left( \tau \right)} \\ \mathtt{m^2}\left( \tau \right) &= \mathtt{V\prime\prime}\left( \tau \right) \\ \mathtt{{\epsilon}s}\left( \tau \right) &= \frac{\left( \frac{\mathtt{V\prime}\left( \tau \right)}{V\left( \tau \right)} \right)^{2}}{16 \pi} \\ \mathtt{{\eta}s}\left( \tau \right) &= \frac{\mathtt{V\prime\prime}\left( \tau \right)}{8 V\left( \tau \right) \pi} \\ \delta\left( \tau \right) \epsilon &= 0 \\ \sigma\left( \tau \right) \epsilon &= 0 \\ \mathtt{c_s^2}\left( \tau \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(Q)

\[ \begin{align*} \end{align*} \]

Curvature (effective)

SymBoltz.curvature — Function

curvature(g; name = :K, kwargs...)

Create a species that effectively accounts for curvature in the spacetime with metric g.

K = SymBoltz.curvature(g)

Equations

equations(K)

\[ \begin{align*} w\left( \tau \right) &= \frac{-1}{3} \\ \rho\left( \tau \right) &= \frac{3 \mathtt{\Omega_0}}{8 \left( a\left( \tau \right) \right)^{2} \pi} \\ P\left( \tau \right) &= w\left( \tau \right) \rho\left( \tau \right) \\ \delta\left( \tau \right) &= 0 \\ \theta\left( \tau \right) &= 0 \\ \mathtt{c_s^2}\left( \tau \right) &= 0 \\ \sigma\left( \tau \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(K)

\[ \begin{align*} \end{align*} \]