Components

Spacetimes

SymBoltz.metric — Function

metric(; name = :g, kwargs...)

Create a symbolic component for the perturbed FLRW spacetime metric in the conformal Newtonian gauge with sign signature diag(-1, +1, +1, +1).

g = SymBoltz.metric()

Equations

equations(g)

\[ \begin{align*} z\left( t \right) &= -1 + \frac{1}{a\left( t \right)} \\ \mathscr{E}\left( t \right) &= \frac{\frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t}}{a\left( t \right)} \\ E\left( t \right) &= \frac{\mathscr{E}\left( t \right)}{a\left( t \right)} \\ \mathscr{H}\left( t \right) &= 3.2408 \cdot 10^{-18} h \mathscr{E}\left( t \right) \\ H\left( t \right) &= 3.2408 \cdot 10^{-18} h E\left( t \right) \\ \mathtt{g_{1 1}}\left( t \right) &= \left( a\left( t \right) \right)^{2} \left( -1 - 2 \Psi\left( t \right) \epsilon \right) \\ \mathtt{g_{2 2}}\left( t \right) &= \left( a\left( t \right) \right)^{2} \left( 1 - 2 \Phi\left( t \right) \epsilon \right) \end{align*} \]

Gravitational theories

SymBoltz.general_relativity — Function

general_relativity(g; name = :G, kwargs...)

Create a symbolic component for the general relativistic (GR) theory of gravity in the spacetime with the metric g.

G = SymBoltz.general_relativity(g)

Equations

equations(G)

\[ \begin{align*} \frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t} &= \sqrt{\frac{8}{3} \left( a\left( t \right) \right)^{4} \rho\left( t \right) \pi} \\ \Delta\left( t \right) &= \frac{\mathrm{d}}{\mathrm{d}t} \frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t} + \frac{ - \left( \frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t} \right)^{2} \left( 1 + \frac{ - 3 P\left( t \right)}{\rho\left( t \right)} \right)}{2 a\left( t \right)} \\ \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \frac{ - k^{2} \Phi\left( t \right)}{3 \mathscr{E}\left( t \right)} + \frac{ - \frac{4}{3} \left( a\left( t \right) \right)^{2} \mathtt{\delta\rho}\left( t \right) \pi}{\mathscr{E}\left( t \right)} - \mathscr{E}\left( t \right) \Psi\left( t \right) \right) \epsilon \\ k^{2} \left( - \Psi\left( t \right) + \Phi\left( t \right) \right) \epsilon &= 12 \left( a\left( t \right) \right)^{2} \Pi\left( t \right) \pi \epsilon \end{align*} \]

Initialization equations

initialization_equations(G)

\[ \begin{align*} \end{align*} \]

SymBoltz.brans_dicke — Function

brans_dicke(g; name = :G, kwargs...)

Create a symbolic component for the Brans-Dicke (BD) theory of gravity in the spacetime with the metric g.

G = SymBoltz.brans_dicke(g)

Equations

equations(G)

\[ \begin{align*} \frac{\mathrm{d}}{\mathrm{d}t} \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} &= \frac{8 \left( a\left( t \right) \right)^{2} \left( \rho\left( t \right) - 3 P\left( t \right) \right) \pi}{3 + 2 \omega} + \frac{ - 2 \frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t} \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t}}{a\left( t \right)} \\ G\left( t \right) &= \frac{4 + 2 \omega}{\phi\left( t \right) \left( 3 + 2 \omega \right)} \\ \frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t} &= \sqrt{\frac{\frac{8}{3} \left( a\left( t \right) \right)^{4} \rho\left( t \right) \pi}{\phi\left( t \right)} + \left( \frac{\frac{1}{2} a\left( t \right) \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t}}{\phi\left( t \right)} \right)^{2} + \frac{1}{6} \left( \frac{a\left( t \right) \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t}}{\phi\left( t \right)} \right)^{2} \omega} + \frac{ - \frac{1}{2} a\left( t \right) \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t}}{\phi\left( t \right)} \\ \Delta\left( t \right) &= \frac{\frac{1}{2} \frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t} \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t}}{\phi\left( t \right)} + \frac{\frac{1}{2} \frac{\mathrm{d}}{\mathrm{d}t} \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} a\left( t \right)}{\phi\left( t \right)} + \frac{ - \left( \frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t} \right)^{2}}{2 a\left( t \right)} + \frac{\mathrm{d}}{\mathrm{d}t} \frac{\mathrm{d} a\left( t \right)}{\mathrm{d}t} + \frac{4 \left( a\left( t \right) \right)^{3} \pi P\left( t \right)}{\phi\left( t \right)} + \frac{1}{4} \left( \frac{\frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t}}{\phi\left( t \right)} \right)^{2} a\left( t \right) \omega \\ \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} \epsilon &= \frac{\left( 3 \frac{\mathrm{d} \mathtt{\delta\phi}\left( t \right)}{\mathrm{d}t} \mathscr{E}\left( t \right) + \left( k^{2} + 3 \left( \mathscr{E}\left( t \right) \right)^{2} \right) \mathtt{\delta\phi}\left( t \right) - 2 k^{2} \phi\left( t \right) \Phi\left( t \right) + \frac{1}{2} \left( - 2 \frac{\mathrm{d} \mathtt{\delta\phi}\left( t \right)}{\mathrm{d}t} \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} + \left( \frac{\frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t}}{\phi\left( t \right)} \right)^{2} \mathtt{\delta\phi}\left( t \right) \right) \omega - 8 \left( a\left( t \right) \right)^{2} \left( \mathtt{\delta\rho}\left( t \right) + 2 \Psi\left( t \right) \rho\left( t \right) \right) \pi \right) \epsilon}{3 \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} + 6 \phi\left( t \right) \mathscr{E}\left( t \right)} \\ \Psi\left( t \right) \epsilon &= \left( \frac{ - 12 \left( a\left( t \right) \right)^{2} \Pi\left( t \right) \pi}{k^{2} \phi\left( t \right)} + \Phi\left( t \right) + \frac{ - \mathtt{\delta\phi}\left( t \right)}{\phi\left( t \right)} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \frac{\mathrm{d} \mathtt{\delta\phi}\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \frac{ - 8 \left( a\left( t \right) \right)^{2} \left( 3 \mathtt{{\delta}P}\left( t \right) - \mathtt{\delta\rho}\left( t \right) \right) \pi}{3 + 2 \omega} - 2 \frac{\mathrm{d} \mathtt{\delta\phi}\left( t \right)}{\mathrm{d}t} \mathscr{E}\left( t \right) + 2 \frac{\mathrm{d}}{\mathrm{d}t} \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} \Psi\left( t \right) + \left( \frac{\mathrm{d} \Psi\left( t \right)}{\mathrm{d}t} + 3 \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} \right) \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} - k^{2} \mathtt{\delta\phi}\left( t \right) + 4 \mathscr{E}\left( t \right) \Psi\left( t \right) \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} \right) \epsilon \end{align*} \]

Initialization equations

initialization_equations(G)

\[ \begin{align*} \mathtt{\delta\phi}\left( t \right) \epsilon &= 0 \\ \frac{\mathrm{d} \mathtt{\delta\phi}\left( t \right)}{\mathrm{d}t} \epsilon &= 0 \end{align*} \]

Species

Generic

SymBoltz.species_constant_eos — Function

species_constant_eos(g, w, cₛ² = w, ẇ = 0, _σ = 0; θinteract = false, kwargs...)

Create a symbolic component for a particle species with equation of state w ~ P/ρ in the spacetime with the metric g.

using SymBoltz: t
@variables w(t)
s = SymBoltz.species_constant_eos(g, w)

Equations

equations(s)

\[ \begin{align*} w\left( t \right) &= w\left( t \right) \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \rho\left( t \right) &= \frac{3 \left( a\left( t \right) \right)^{ - 3 \left( 1 + w\left( t \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( t \right) &= \frac{8}{3} \rho\left( t \right) \pi \\ \frac{\mathrm{d} \delta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \left( \theta\left( t \right) - 3 \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} \right) \left( -1 - w\left( t \right) \right) - 3 \mathscr{E}\left( t \right) \left( - w\left( t \right) + \mathtt{c_s^2}\left( t \right) \right) \delta\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \mathtt{{\theta}interaction}\left( t \right) + \frac{k^{2} \mathtt{c_s^2}\left( t \right) \delta\left( t \right)}{1 + w\left( t \right)} - k^{2} \sigma\left( t \right) + k^{2} \Psi\left( t \right) - \mathscr{E}\left( t \right) \theta\left( t \right) \left( 1 - 3 w\left( t \right) \right) \right) \epsilon \\ u\left( t \right) \epsilon &= \frac{\theta\left( t \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( t \right) \epsilon &= \frac{\mathrm{d} u\left( t \right)}{\mathrm{d}t} \epsilon \\ \Delta\left( t \right) \epsilon &= \left( \frac{3 \mathscr{E}\left( t \right) \left( 1 + w\left( t \right) \right)}{\theta\left( t \right)} + \delta\left( t \right) \right) \epsilon \\ \sigma\left( t \right) \epsilon &= 0 \\ \mathtt{{\theta}interaction}\left( t \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(s)

\[ \begin{align*} \delta\left( t \right) \epsilon &= - \frac{3}{2} \Psi\left( t \right) \left( 1 + w\left( t \right) \right) \epsilon \\ \theta\left( t \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right)} \end{align*} \]

Relativistic

SymBoltz.radiation — Function

radiation(g; name = :r, kwargs...)

Create a particle species for radiation (with equation of state w ~ 1/3) in the spacetime with metric g.

r = SymBoltz.radiation(g)

Equations

equations(r)

\[ \begin{align*} T\left( t \right) &= \frac{\mathtt{T{_0}}}{a\left( t \right)} \\ w\left( t \right) &= \frac{1}{3} \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \rho\left( t \right) &= \frac{3 \left( a\left( t \right) \right)^{ - 3 \left( 1 + w\left( t \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( t \right) &= \frac{8}{3} \rho\left( t \right) \pi \\ \frac{\mathrm{d} \delta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \left( \theta\left( t \right) - 3 \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} \right) \left( -1 - w\left( t \right) \right) - 3 \mathscr{E}\left( t \right) \left( - w\left( t \right) + \mathtt{c_s^2}\left( t \right) \right) \delta\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \mathtt{{\theta}interaction}\left( t \right) + \frac{k^{2} \mathtt{c_s^2}\left( t \right) \delta\left( t \right)}{1 + w\left( t \right)} - k^{2} \sigma\left( t \right) + k^{2} \Psi\left( t \right) - \mathscr{E}\left( t \right) \theta\left( t \right) \left( 1 - 3 w\left( t \right) \right) \right) \epsilon \\ u\left( t \right) \epsilon &= \frac{\theta\left( t \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( t \right) \epsilon &= \frac{\mathrm{d} u\left( t \right)}{\mathrm{d}t} \epsilon \\ \Delta\left( t \right) \epsilon &= \left( \frac{3 \mathscr{E}\left( t \right) \left( 1 + w\left( t \right) \right)}{\theta\left( t \right)} + \delta\left( t \right) \right) \epsilon \\ \sigma\left( t \right) \epsilon &= 0 \\ \mathtt{{\theta}interaction}\left( t \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(r)

\[ \begin{align*} \delta\left( t \right) \epsilon &= - \frac{3}{2} \Psi\left( t \right) \left( 1 + w\left( t \right) \right) \epsilon \\ \theta\left( t \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right)} \end{align*} \]

SymBoltz.photons — Function

photons(g; polarization = true, lmax = 6, name = :γ, kwargs...)

Create a particle species for photons in the spacetime with metric g.

γ = SymBoltz.photons(g; lmax = 6)

Equations

equations(γ)

\[ \begin{align*} \frac{\mathrm{d} F\left( t \right)_{0}}{\mathrm{d}t} \epsilon &= \left( 4 \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} - F\left( t \right)_{1} k \right) \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{1}}{\mathrm{d}t} \epsilon &= \left( \frac{ - \frac{4}{3} \mathtt{\dot{\tau}}\left( t \right) \left( - \theta\left( t \right) + \mathtt{{\theta}b}\left( t \right) \right)}{k} + \frac{1}{3} \left( F\left( t \right)_{0} - 2 F\left( t \right)_{2} + 4 \Psi\left( t \right) \right) k \right) \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{2}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{5} \left( 2 F\left( t \right)_{1} - 3 F\left( t \right)_{3} \right) k + \frac{9}{10} F\left( t \right)_{2} \mathtt{\dot{\tau}}\left( t \right) - \frac{1}{10} \left( G\left( t \right)_{0} + G\left( t \right)_{2} \right) \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{3}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{7} \left( 3 F\left( t \right)_{2} - 4 F\left( t \right)_{4} \right) k + F\left( t \right)_{3} \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{4}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{9} \left( 4 F\left( t \right)_{3} - 5 F\left( t \right)_{5} \right) k + F\left( t \right)_{4} \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{5}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{11} \left( 5 F\left( t \right)_{4} - 6 F\left( t \right)_{6} \right) k + F\left( t \right)_{5} \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{6}}{\mathrm{d}t} \epsilon &= \left( F\left( t \right)_{5} k + F\left( t \right)_{6} \mathtt{\dot{\tau}}\left( t \right) - 7 F\left( t \right)_{6} \mathscr{E}\left( t \right) \right) \epsilon \\ \delta\left( t \right) \epsilon &= F\left( t \right)_{0} \epsilon \\ \theta\left( t \right) \epsilon &= \frac{3}{4} F\left( t \right)_{1} k \epsilon \\ \sigma\left( t \right) \epsilon &= \frac{1}{2} F\left( t \right)_{2} \epsilon \\ \Pi\left( t \right) \epsilon &= \left( F\left( t \right)_{2} + G\left( t \right)_{0} + G\left( t \right)_{2} \right) \epsilon \\ \mathtt{\dot{\Pi}}\left( t \right) \epsilon &= \frac{\mathrm{d} \Pi\left( t \right)}{\mathrm{d}t} \epsilon \\ \mathtt{c_s^2}\left( t \right) \epsilon &= \frac{1}{3} \epsilon \\ \Theta\left( t \right)_{0} \epsilon &= \frac{1}{4} F\left( t \right)_{0} \epsilon \\ \Theta\left( t \right)_{1} \epsilon &= \frac{1}{4} F\left( t \right)_{1} \epsilon \\ \Theta\left( t \right)_{2} \epsilon &= \frac{1}{4} F\left( t \right)_{2} \epsilon \\ \Theta\left( t \right)_{3} \epsilon &= \frac{1}{4} F\left( t \right)_{3} \epsilon \\ \Theta\left( t \right)_{4} \epsilon &= \frac{1}{4} F\left( t \right)_{4} \epsilon \\ \Theta\left( t \right)_{5} \epsilon &= \frac{1}{4} F\left( t \right)_{5} \epsilon \\ \Theta\left( t \right)_{6} \epsilon &= \frac{1}{4} F\left( t \right)_{6} \epsilon \\ \frac{\mathrm{d} G\left( t \right)_{0}}{\mathrm{d}t} \epsilon &= \left( \left( G\left( t \right)_{0} - \frac{1}{2} \Pi\left( t \right) \right) \mathtt{\dot{\tau}}\left( t \right) - G\left( t \right)_{1} k \right) \epsilon \\ \frac{\mathrm{d} G\left( t \right)_{1}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{3} \left( G\left( t \right)_{0} - 2 G\left( t \right)_{2} \right) k + G\left( t \right)_{1} \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} G\left( t \right)_{2}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{5} \left( 2 G\left( t \right)_{1} - 3 G\left( t \right)_{3} \right) k + \left( G\left( t \right)_{2} - \frac{1}{10} \Pi\left( t \right) \right) \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} G\left( t \right)_{3}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{7} \left( 3 G\left( t \right)_{2} - 4 G\left( t \right)_{4} \right) k + G\left( t \right)_{3} \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} G\left( t \right)_{4}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{9} \left( 4 G\left( t \right)_{3} - 5 G\left( t \right)_{5} \right) k + G\left( t \right)_{4} \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} G\left( t \right)_{5}}{\mathrm{d}t} \epsilon &= \left( \frac{1}{11} \left( 5 G\left( t \right)_{4} - 6 G\left( t \right)_{6} \right) k + G\left( t \right)_{5} \mathtt{\dot{\tau}}\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} G\left( t \right)_{6}}{\mathrm{d}t} \epsilon &= \left( G\left( t \right)_{5} k + G\left( t \right)_{6} \mathtt{\dot{\tau}}\left( t \right) - 7 G\left( t \right)_{6} \mathscr{E}\left( t \right) \right) \epsilon \\ T\left( t \right) &= \frac{\mathtt{T{_0}}}{a\left( t \right)} \\ w\left( t \right) &= \frac{1}{3} \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \rho\left( t \right) &= 0.11937 \left( a\left( t \right) \right)^{ - 3 \left( 1 + w\left( t \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( t \right) &= 8.3776 \rho\left( t \right) \end{align*} \]

Initialization equations

initialization_equations(γ)

\[ \begin{align*} F\left( t \right)_{0} \epsilon &= - 2 \Psi\left( t \right) \epsilon \\ F\left( t \right)_{1} \epsilon &= \frac{\frac{2}{3} k \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right)} \\ F\left( t \right)_{2} \epsilon &= \frac{ - \frac{8}{15} F\left( t \right)_{1} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ F\left( t \right)_{3} \epsilon &= \frac{ - \frac{3}{7} F\left( t \right)_{2} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ F\left( t \right)_{4} \epsilon &= \frac{ - \frac{4}{9} F\left( t \right)_{3} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ F\left( t \right)_{5} \epsilon &= \frac{ - \frac{5}{11} F\left( t \right)_{4} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ F\left( t \right)_{6} \epsilon &= \frac{ - \frac{6}{13} F\left( t \right)_{5} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ G\left( t \right)_{0} \epsilon &= \frac{5}{16} F\left( t \right)_{2} \epsilon \\ G\left( t \right)_{1} \epsilon &= \frac{ - \frac{1}{16} F\left( t \right)_{2} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ G\left( t \right)_{2} \epsilon &= \frac{1}{16} F\left( t \right)_{2} \epsilon \\ G\left( t \right)_{3} \epsilon &= \frac{ - \frac{3}{7} G\left( t \right)_{2} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ G\left( t \right)_{4} \epsilon &= \frac{ - \frac{4}{9} G\left( t \right)_{3} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ G\left( t \right)_{5} \epsilon &= \frac{ - \frac{5}{11} G\left( t \right)_{4} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \\ G\left( t \right)_{6} \epsilon &= \frac{ - \frac{6}{13} G\left( t \right)_{5} k \epsilon}{\mathtt{\dot{\tau}}\left( t \right)} \end{align*} \]

Non-relativistic

SymBoltz.matter — Function

matter(g; name = :m, kwargs...)

Create a particle species for matter (with equation of state w ~ 0) in the spacetime with metric g.

m = SymBoltz.matter(g)

Equations

equations(m)

\[ \begin{align*} w\left( t \right) &= 0 \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \rho\left( t \right) &= \frac{3 \left( a\left( t \right) \right)^{ - 3 \left( 1 + w\left( t \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( t \right) &= \frac{8}{3} \rho\left( t \right) \pi \\ \frac{\mathrm{d} \delta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \left( \theta\left( t \right) - 3 \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} \right) \left( -1 - w\left( t \right) \right) - 3 \mathscr{E}\left( t \right) \left( - w\left( t \right) + \mathtt{c_s^2}\left( t \right) \right) \delta\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \mathtt{{\theta}interaction}\left( t \right) + \frac{k^{2} \mathtt{c_s^2}\left( t \right) \delta\left( t \right)}{1 + w\left( t \right)} - k^{2} \sigma\left( t \right) + k^{2} \Psi\left( t \right) - \mathscr{E}\left( t \right) \theta\left( t \right) \left( 1 - 3 w\left( t \right) \right) \right) \epsilon \\ u\left( t \right) \epsilon &= \frac{\theta\left( t \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( t \right) \epsilon &= \frac{\mathrm{d} u\left( t \right)}{\mathrm{d}t} \epsilon \\ \Delta\left( t \right) \epsilon &= \left( \frac{3 \mathscr{E}\left( t \right) \left( 1 + w\left( t \right) \right)}{\theta\left( t \right)} + \delta\left( t \right) \right) \epsilon \\ \sigma\left( t \right) \epsilon &= 0 \\ \mathtt{{\theta}interaction}\left( t \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(m)

\[ \begin{align*} \delta\left( t \right) \epsilon &= - \frac{3}{2} \Psi\left( t \right) \left( 1 + w\left( t \right) \right) \epsilon \\ \theta\left( t \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right)} \end{align*} \]

SymBoltz.cold_dark_matter — Function

cold_dark_matter(g; name = :c, kwargs...)

Create a particle species for cold dark matter in the spacetime with metric g.

c = SymBoltz.cold_dark_matter(g)

Equations

equations(c)

\[ \begin{align*} \mathtt{c_s^2}\left( t \right) \epsilon &= 0 \\ w\left( t \right) &= 0 \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \rho\left( t \right) &= \frac{3 \left( a\left( t \right) \right)^{ - 3 \left( 1 + w\left( t \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( t \right) &= \frac{8}{3} \rho\left( t \right) \pi \\ \frac{\mathrm{d} \delta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \left( \theta\left( t \right) - 3 \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} \right) \left( -1 - w\left( t \right) \right) - 3 \mathscr{E}\left( t \right) \left( - w\left( t \right) + \mathtt{c_s^2}\left( t \right) \right) \delta\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \mathtt{{\theta}interaction}\left( t \right) + \frac{k^{2} \mathtt{c_s^2}\left( t \right) \delta\left( t \right)}{1 + w\left( t \right)} - k^{2} \sigma\left( t \right) + k^{2} \Psi\left( t \right) - \mathscr{E}\left( t \right) \theta\left( t \right) \left( 1 - 3 w\left( t \right) \right) \right) \epsilon \\ u\left( t \right) \epsilon &= \frac{\theta\left( t \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( t \right) \epsilon &= \frac{\mathrm{d} u\left( t \right)}{\mathrm{d}t} \epsilon \\ \Delta\left( t \right) \epsilon &= \left( \frac{3 \mathscr{E}\left( t \right) \left( 1 + w\left( t \right) \right)}{\theta\left( t \right)} + \delta\left( t \right) \right) \epsilon \\ \sigma\left( t \right) \epsilon &= 0 \\ \mathtt{{\theta}interaction}\left( t \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(c)

\[ \begin{align*} \delta\left( t \right) \epsilon &= - \frac{3}{2} \Psi\left( t \right) \left( 1 + w\left( t \right) \right) \epsilon \\ \theta\left( t \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right)} \end{align*} \]

SymBoltz.baryons — Function

baryons(g; recombination=true, name = :b, kwargs...)

Create a particle species for baryons in the spacetime with metric g.

b = SymBoltz.baryons(g)

Equations

equations(b)

\[ \begin{align*} \mathtt{c_s^2}\left( t \right) \epsilon &= \mathtt{rec.c_s^2}\left( t \right) \epsilon \\ w\left( t \right) &= 0 \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \rho\left( t \right) &= \frac{3 \left( a\left( t \right) \right)^{ - 3 \left( 1 + w\left( t \right) \right)} \mathtt{\Omega_0}}{8 \pi} \\ \Omega\left( t \right) &= \frac{8}{3} \rho\left( t \right) \pi \\ \frac{\mathrm{d} \delta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \left( \theta\left( t \right) - 3 \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} \right) \left( -1 - w\left( t \right) \right) - 3 \mathscr{E}\left( t \right) \left( - w\left( t \right) + \mathtt{c_s^2}\left( t \right) \right) \delta\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \mathtt{{\theta}interaction}\left( t \right) + \frac{k^{2} \mathtt{c_s^2}\left( t \right) \delta\left( t \right)}{1 + w\left( t \right)} - k^{2} \sigma\left( t \right) + k^{2} \Psi\left( t \right) - \mathscr{E}\left( t \right) \theta\left( t \right) \left( 1 - 3 w\left( t \right) \right) \right) \epsilon \\ u\left( t \right) \epsilon &= \frac{\theta\left( t \right) \epsilon}{k} \\ \mathtt{\dot{u}}\left( t \right) \epsilon &= \frac{\mathrm{d} u\left( t \right)}{\mathrm{d}t} \epsilon \\ \Delta\left( t \right) \epsilon &= \left( \frac{3 \mathscr{E}\left( t \right) \left( 1 + w\left( t \right) \right)}{\theta\left( t \right)} + \delta\left( t \right) \right) \epsilon \\ \sigma\left( t \right) \epsilon &= 0 \\ \mathtt{rec.nH}\left( t \right) &= 5.9743 \cdot 10^{26} \left( 1 - \mathtt{rec.Yp} \right) \mathtt{rec.{\rho}b}\left( t \right) \\ \mathtt{rec.nHe}\left( t \right) &= \mathtt{rec.fHe} \mathtt{rec.nH}\left( t \right) \\ \mathtt{rec.DTb}\left( t \right) &= \frac{ - 0.00015165 \left( \mathtt{rec.T\gamma}\left( t \right) \right)^{4} \mathtt{rec.Xe}\left( t \right) a\left( t \right) \mathtt{rec.{\Delta}T}\left( t \right)}{h \left( 1 + \mathtt{rec.fHe} + \mathtt{rec.Xe}\left( t \right) \right)} - 2 \mathscr{E}\left( t \right) \mathtt{rec.Tb}\left( t \right) \\ \mathtt{rec.DT\gamma}\left( t \right) &= \frac{\mathrm{d} \mathtt{rec.T\gamma}\left( t \right)}{\mathrm{d}t} \\ \frac{\mathrm{d} \mathtt{rec.{\Delta}T}\left( t \right)}{\mathrm{d}t} &= \mathtt{rec.DTb}\left( t \right) - \mathtt{rec.DT\gamma}\left( t \right) \\ \mathtt{rec.Tb}\left( t \right) &= \mathtt{rec.T\gamma}\left( t \right) + \mathtt{rec.{\Delta}T}\left( t \right) \\ \mathtt{rec.{\beta}b}\left( t \right) &= \frac{1}{1.3806 \cdot 10^{-23} \mathtt{rec.Tb}\left( t \right)} \\ \mathtt{rec.{\lambda}e}\left( t \right) &= \frac{6.6261 \cdot 10^{-34}}{\sqrt{\frac{5.7236 \cdot 10^{-30}}{\mathtt{rec.{\beta}b}\left( t \right)}}} \\ \mathtt{rec.\mu}\left( t \right) &= \frac{1.6738 \cdot 10^{-27}}{1 - 0.74816 \mathtt{rec.Yp} + \left( 1 - \mathtt{rec.Yp} \right) \mathtt{rec.Xe}\left( t \right)} \\ \mathtt{rec.c_s^2}\left( t \right) &= \frac{1.3806 \cdot 10^{-23} \left( \mathtt{rec.Tb}\left( t \right) + \frac{ - \frac{\mathrm{d} \mathtt{rec.Tb}\left( t \right)}{\mathrm{d}t}}{3 \mathscr{E}\left( t \right)} \right)}{8.9876 \cdot 10^{16} \mathtt{rec.\mu}\left( t \right)} \\ \mathtt{rec.{\alpha}H}\left( t \right) &= \frac{4.9123 \cdot 10^{-19}}{0.0034166 \left( \mathtt{rec.Tb}\left( t \right) \right)^{0.6166} \left( 1 + 0.0050847 \left( \mathtt{rec.Tb}\left( t \right) \right)^{0.53} \right)} \\ \mathtt{rec.{\beta}H}\left( t \right) &= \frac{e^{ - 5.4468 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( t \right)} \mathtt{rec.{\alpha}H}\left( t \right)}{\left( \mathtt{rec.{\lambda}e}\left( t \right) \right)^{3}} \\ \mathtt{rec.KH0}\left( t \right) &= \frac{1.7966 \cdot 10^{-21}}{25.133 H\left( t \right)} \\ \mathtt{rec.KH1}\left( t \right) &= \left( - 0.14 e^{ - 30.864 \left( 7.28 + \log\left( a\left( t \right) \right) \right)^{2}} + 0.079 e^{ - 9.1827 \left( 6.73 + \log\left( a\left( t \right) \right) \right)^{2}} \right) \mathtt{rec.KH0}\left( t \right) \\ \mathtt{rec.KH}\left( t \right) &= \mathtt{rec.KH1}\left( t \right) + \mathtt{rec.KH0}\left( t \right) \\ \mathtt{rec.CH}\left( t \right) &= \frac{1 + 8.2246 \left( 1 - \mathtt{rec.XH^+}\left( t \right) \right) \mathtt{rec.KH}\left( t \right) \mathtt{rec.nH}\left( t \right)}{1 + \left( 1 - \mathtt{rec.XH^+}\left( t \right) \right) \mathtt{rec.KH}\left( t \right) \left( 8.2246 + \mathtt{rec.{\beta}H}\left( t \right) \right) \mathtt{rec.nH}\left( t \right)} \\ \frac{\mathrm{d} \mathtt{rec.XH^+}\left( t \right)}{\mathrm{d}t} &= \frac{ - \left( \mathtt{rec.ne}\left( t \right) \mathtt{rec.{\alpha}H}\left( t \right) \mathtt{rec.XH^+}\left( t \right) - e^{ - 1.634 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( t \right)} \left( 1 - \mathtt{rec.XH^+}\left( t \right) \right) \mathtt{rec.{\beta}H}\left( t \right) \right) \mathtt{rec.CH}\left( t \right) a\left( t \right)}{3.2408 \cdot 10^{-18} h} \\ \mathtt{rec.{\alpha}He}\left( t \right) &= \frac{1.803 \cdot 10^{-17}}{\left( 1 + \sqrt{7.6913 \cdot 10^{-6} \mathtt{rec.Tb}\left( t \right)} \right)^{1.711} \left( 1 + \sqrt{0.33333 \mathtt{rec.Tb}\left( t \right)} \right)^{0.289} \sqrt{0.33333 \mathtt{rec.Tb}\left( t \right)}} \\ \mathtt{rec.{\beta}He}\left( t \right) &= \frac{4 e^{ - 6.3633 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( t \right)} \mathtt{rec.{\alpha}He}\left( t \right)}{\left( \mathtt{rec.{\lambda}e}\left( t \right) \right)^{3}} \\ \mathtt{rec.KHe0^{- 1}}\left( t \right) &= 1.2597 \cdot 10^{23} H\left( t \right) \\ \mathtt{rec.KHe1^{- 1}}\left( t \right) &= - \mathtt{rec.KHe0^{- 1}}\left( t \right) e^{\frac{ - 5.3949 \cdot 10^{9} \mathtt{rec.nHe}\left( t \right) \left( 1 - \mathtt{rec.XHe^+}\left( t \right) \right)}{\mathtt{rec.KHe0^{- 1}}\left( t \right)}} \\ \mathtt{rec.{\gamma}2Ps}\left( t \right) &= \left|\frac{4.8487 \cdot 10^{26} \mathtt{rec.fHe} \left( 1 - \mathtt{rec.XHe^+}\left( t \right) \right)}{4.8748 \cdot 10^{26} \left( 1 - \mathtt{rec.XH^+}\left( t \right) \right) \sqrt{\frac{6.2832}{5.9736 \cdot 10^{-10} \mathtt{rec.{\beta}b}\left( t \right)}}}\right| \\ \mathtt{rec.KHe2^{- 1}}\left( t \right) &= \frac{5.3949 \cdot 10^{9} \mathtt{rec.nHe}\left( t \right) \left( 1 - \mathtt{rec.XHe^+}\left( t \right) \right)}{1 + 0.36 \left( \mathtt{rec.{\gamma}2Ps}\left( t \right) \right)^{0.86}} \\ \mathtt{rec.KHe}\left( t \right) &= \frac{1}{\mathtt{rec.KHe2^{- 1}}\left( t \right) + \mathtt{rec.KHe0^{- 1}}\left( t \right) + \mathtt{rec.KHe1^{- 1}}\left( t \right)} \\ \mathtt{rec.CHe}\left( t \right) &= \frac{e^{ - 9.6491 \cdot 10^{-20} \mathtt{rec.{\beta}b}\left( t \right)} + 51.3 \mathtt{rec.nHe}\left( t \right) \mathtt{rec.KHe}\left( t \right) \left( 1 - \mathtt{rec.XHe^+}\left( t \right) \right)}{e^{ - 9.6491 \cdot 10^{-20} \mathtt{rec.{\beta}b}\left( t \right)} + \mathtt{rec.nHe}\left( t \right) \mathtt{rec.KHe}\left( t \right) \left( 51.3 + \mathtt{rec.{\beta}He}\left( t \right) \right) \left( 1 - \mathtt{rec.XHe^+}\left( t \right) \right)} \\ \mathtt{rec.DXHe^+\_singlet}\left( t \right) &= \frac{\left( - \mathtt{rec.ne}\left( t \right) \mathtt{rec.{\alpha}He}\left( t \right) \mathtt{rec.XHe^+}\left( t \right) - \mathtt{rec.{\beta}He}\left( t \right) e^{ - 3.303 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( t \right)} \left( -1 + \mathtt{rec.XHe^+}\left( t \right) \right) \right) \mathtt{rec.CHe}\left( t \right) a\left( t \right)}{3.2408 \cdot 10^{-18} h} \\ \mathtt{rec.{\alpha}He3}\left( t \right) &= \frac{4.9431 \cdot 10^{-17}}{\left( 1 + \sqrt{7.6913 \cdot 10^{-6} \mathtt{rec.Tb}\left( t \right)} \right)^{1.761} \left( 1 + \sqrt{0.33333 \mathtt{rec.Tb}\left( t \right)} \right)^{0.239} \sqrt{0.33333 \mathtt{rec.Tb}\left( t \right)}} \\ \mathtt{rec.{\beta}He3}\left( t \right) &= \frac{1.3333 \mathtt{rec.{\alpha}He3}\left( t \right) e^{ - 7.6388 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( t \right)}}{\left( \mathtt{rec.{\lambda}e}\left( t \right) \right)^{3}} \\ \mathtt{rec.{\tau}He3}\left( t \right) &= \frac{1.102 \cdot 10^{-19} \mathtt{rec.nHe}\left( t \right) \left( 1 - \mathtt{rec.XHe^+}\left( t \right) \right)}{25.133 H\left( t \right)} \\ \mathtt{rec.pHe3}\left( t \right) &= \frac{1 - e^{ - \mathtt{rec.{\tau}He3}\left( t \right)}}{\mathtt{rec.{\tau}He3}\left( t \right)} \\ \mathtt{rec.{\gamma}2Pt}\left( t \right) &= \left|\frac{1.8633 \cdot 10^{-12} \mathtt{rec.fHe} \left( 1 - \mathtt{rec.XHe^+}\left( t \right) \right)}{1.8917 \cdot 10^{-5} \left( 1 - \mathtt{rec.XH^+}\left( t \right) \right) \sqrt{\frac{6.2832}{5.9736 \cdot 10^{-10} \mathtt{rec.{\beta}b}\left( t \right)}}}\right| \\ \mathtt{rec.CHe3}\left( t \right) &= \frac{1 \cdot 10^{-10} + 177.58 \left( \frac{\frac{1}{3}}{1 + 0.66 \left( \mathtt{rec.{\gamma}2Pt}\left( t \right) \right)^{0.9}} + \mathtt{rec.pHe3}\left( t \right) \right) e^{ - 1.8337 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( t \right)}}{1 \cdot 10^{-10} + \mathtt{rec.{\beta}He3}\left( t \right) + 177.58 \left( \frac{\frac{1}{3}}{1 + 0.66 \left( \mathtt{rec.{\gamma}2Pt}\left( t \right) \right)^{0.9}} + \mathtt{rec.pHe3}\left( t \right) \right) e^{ - 1.8337 \cdot 10^{-19} \mathtt{rec.{\beta}b}\left( t \right)}} \\ \mathtt{rec.DXHe^+\_triplet}\left( t \right) &= \frac{ - \left( \mathtt{rec.{\alpha}He3}\left( t \right) \mathtt{rec.ne}\left( t \right) \mathtt{rec.XHe^+}\left( t \right) - 3 e^{ - 3.1755 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( t \right)} \mathtt{rec.{\beta}He3}\left( t \right) \left( 1 - \mathtt{rec.XHe^+}\left( t \right) \right) \right) a\left( t \right) \mathtt{rec.CHe3}\left( t \right)}{3.2408 \cdot 10^{-18} h} \\ \frac{\mathrm{d} \mathtt{rec.XHe^+}\left( t \right)}{\mathrm{d}t} &= \mathtt{rec.DXHe^+\_triplet}\left( t \right) + \mathtt{rec.DXHe^+\_singlet}\left( t \right) \\ \mathtt{rec.RHe^+}\left( t \right) &= \frac{e^{ - 8.7187 \cdot 10^{-18} \mathtt{rec.{\beta}b}\left( t \right)}}{\left( \mathtt{rec.{\lambda}e}\left( t \right) \right)^{3} \mathtt{rec.nH}\left( t \right)} \\ \mathtt{rec.XHe^{+ +}}\left( t \right) &= \frac{2 \mathtt{rec.fHe} \mathtt{rec.RHe^+}\left( t \right)}{\left( 1 + \mathtt{rec.fHe} + \mathtt{rec.RHe^+}\left( t \right) \right) \left( 1 + \sqrt{1 + \frac{4 \mathtt{rec.fHe} \mathtt{rec.RHe^+}\left( t \right)}{\left( 1 + \mathtt{rec.fHe} + \mathtt{rec.RHe^+}\left( t \right) \right)^{2}}} \right)} \\ \mathtt{rec.re1Xe}\left( t \right) &= 0.5 \left( 1 + \mathtt{rec.fHe} + \left( 1 + \mathtt{rec.fHe} \right) \tanh\left( \frac{\left( 1 + \mathtt{rec.re1z} \right)^{1.5} - \left( 1 + z\left( t \right) \right)^{1.5}}{0.75 \left( 1 + \mathtt{rec.re1z} \right)^{0.5}} \right) \right) \\ \mathtt{rec.re2Xe}\left( t \right) &= 0.5 \left( \mathtt{rec.fHe} + \mathtt{rec.fHe} \tanh\left( \frac{\left( 1 + \mathtt{rec.re2z} \right)^{1.5} - \left( 1 + z\left( t \right) \right)^{1.5}}{0.75 \left( 1 + \mathtt{rec.re2z} \right)^{0.5}} \right) \right) \\ \mathtt{rec.Xe}\left( t \right) &= \mathtt{rec.re2Xe}\left( t \right) + \mathtt{rec.XH^+}\left( t \right) + \mathtt{rec.re1Xe}\left( t \right) + \mathtt{rec.XHe^{+ +}}\left( t \right) + \mathtt{rec.fHe} \mathtt{rec.XHe^+}\left( t \right) \\ \mathtt{rec.ne}\left( t \right) &= \mathtt{rec.Xe}\left( t \right) \mathtt{rec.nH}\left( t \right) \\ \mathtt{rec.\dot{\tau}}\left( t \right) &= \frac{ - 1.9944 \cdot 10^{-20} \mathtt{rec.ne}\left( t \right) a\left( t \right)}{3.2408 \cdot 10^{-18} h} \\ \frac{\mathrm{d} \mathtt{rec.\tau}\left( t \right)}{\mathrm{d}t} &= \mathtt{rec.\dot{\tau}}\left( t \right) \\ \mathtt{rec.v}\left( t \right) &= \frac{\mathrm{d} e^{ - \mathtt{rec.\tau}\left( t \right)}}{\mathrm{d}t} \\ \mathtt{rec.\dot{v}}\left( t \right) &= \frac{\mathrm{d} \mathtt{rec.v}\left( t \right)}{\mathrm{d}t} \\ \mathtt{rec.\textnormal{\"{v}}}\left( t \right) &= 0 \\ \mathtt{rec.\ddot{\tau}}\left( t \right) &= \frac{\mathrm{d} \mathtt{rec.\dot{\tau}}\left( t \right)}{\mathrm{d}t} \\ \mathtt{rec.\dddot{\tau}}\left( t \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(b)

\[ \begin{align*} \delta\left( t \right) \epsilon &= - \frac{3}{2} \Psi\left( t \right) \left( 1 + w\left( t \right) \right) \epsilon \\ \theta\left( t \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right)} \end{align*} \]

Neutrinos

SymBoltz.massless_neutrinos — Function

massless_neutrinos(g; lmax = 6, name = :ν, kwargs...)

Create a particle species for massless neutrinos in the spacetime with metric g.

ν = SymBoltz.massless_neutrinos(g; lmax = 6)

Equations

equations(ν)

\[ \begin{align*} \frac{\mathrm{d} F\left( t \right)_{0}}{\mathrm{d}t} \epsilon &= \left( 4 \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} - F\left( t \right)_{1} k \right) \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{1}}{\mathrm{d}t} \epsilon &= \frac{1}{3} \left( F\left( t \right)_{0} - 2 F\left( t \right)_{2} + 4 \Psi\left( t \right) \right) k \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{2}}{\mathrm{d}t} \epsilon &= \frac{1}{5} \left( 2 F\left( t \right)_{1} - 3 F\left( t \right)_{3} \right) k \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{3}}{\mathrm{d}t} \epsilon &= \frac{1}{7} \left( 3 F\left( t \right)_{2} - 4 F\left( t \right)_{4} \right) k \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{4}}{\mathrm{d}t} \epsilon &= \frac{1}{9} \left( 4 F\left( t \right)_{3} - 5 F\left( t \right)_{5} \right) k \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{5}}{\mathrm{d}t} \epsilon &= \frac{1}{11} \left( 5 F\left( t \right)_{4} - 6 F\left( t \right)_{6} \right) k \epsilon \\ \frac{\mathrm{d} F\left( t \right)_{6}}{\mathrm{d}t} \epsilon &= \left( F\left( t \right)_{5} k - 7 F\left( t \right)_{6} \mathscr{E}\left( t \right) \right) \epsilon \\ \delta\left( t \right) \epsilon &= F\left( t \right)_{0} \epsilon \\ \theta\left( t \right) \epsilon &= \frac{3}{4} F\left( t \right)_{1} k \epsilon \\ \sigma\left( t \right) \epsilon &= \frac{1}{2} F\left( t \right)_{2} \epsilon \\ \mathtt{c_s^2}\left( t \right) \epsilon &= \frac{1}{3} \epsilon \\ T\left( t \right) &= \frac{\mathtt{T{_0}}}{a\left( t \right)} \\ w\left( t \right) &= \frac{1}{3} \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \rho\left( t \right) &= 0.11937 \left( a\left( t \right) \right)^{ - 3 \left( 1 + w\left( t \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( t \right) &= 8.3776 \rho\left( t \right) \end{align*} \]

Initialization equations

initialization_equations(ν)

\[ \begin{align*} \delta\left( t \right) \epsilon &= - 2 \Psi\left( t \right) \epsilon \\ \theta\left( t \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right)} \\ \sigma\left( t \right) \epsilon &= \frac{1}{15} \left( \frac{k}{\mathscr{E}\left( t \right)} \right)^{2} \Psi\left( t \right) \epsilon \\ F\left( t \right)_{3} \epsilon &= \frac{\frac{3}{7} F\left( t \right)_{2} k \epsilon}{\mathscr{E}\left( t \right)} \\ F\left( t \right)_{4} \epsilon &= \frac{\frac{4}{9} F\left( t \right)_{3} k \epsilon}{\mathscr{E}\left( t \right)} \\ F\left( t \right)_{5} \epsilon &= \frac{\frac{5}{11} F\left( t \right)_{4} k \epsilon}{\mathscr{E}\left( t \right)} \\ F\left( t \right)_{6} \epsilon &= \frac{\frac{6}{13} F\left( t \right)_{5} k \epsilon}{\mathscr{E}\left( t \right)} \end{align*} \]

SymBoltz.massive_neutrinos — Function

massive_neutrinos(g; nx = 5, lmax = 4, name = :h, kwargs...)

Create a particle species for massive neutrinos in the spacetime with metric g.

h = SymBoltz.massive_neutrinos(g; nx=4, lmax = 5)

Equations

equations(h)

\[ \begin{align*} T\left( t \right) &= \frac{\mathtt{T{_0}}}{a\left( t \right)} \\ y\left( t \right) &= \frac{8.9876 \cdot 10^{16} m}{1.3806 \cdot 10^{-23} T\left( t \right)} \\ \mathtt{In}\left( t \right) &= W_{1} + W_{2} + W_{3} + W_{4} \\ \mathtt{I\rho}\left( t \right) &= E\left( t \right)_{1} W_{1} + E\left( t \right)_{2} W_{2} + E\left( t \right)_{3} W_{3} + E\left( t \right)_{4} W_{4} \\ \mathtt{IP}\left( t \right) &= \frac{\left( x_{4} \right)^{2} W_{4}}{E\left( t \right)_{4}} + \frac{\left( x_{1} \right)^{2} W_{1}}{E\left( t \right)_{1}} + \frac{\left( x_{2} \right)^{2} W_{2}}{E\left( t \right)_{2}} + \frac{\left( x_{3} \right)^{2} W_{3}}{E\left( t \right)_{3}} \\ \rho\left( t \right) &= \frac{1.165 \cdot 10^{-16} \left( T\left( t \right) \right)^{4} \mathtt{I\rho}\left( t \right)}{1.4143 \cdot 10^{-8} h^{2}} \\ P\left( t \right) &= \frac{3.8835 \cdot 10^{-17} \left( T\left( t \right) \right)^{4} \mathtt{IP}\left( t \right)}{1.4143 \cdot 10^{-8} h^{2}} \\ w\left( t \right) &= \frac{P\left( t \right)}{\rho\left( t \right)} \\ \Omega\left( t \right) &= \frac{8}{3} \rho\left( t \right) \pi \\ E\left( t \right)_{1} &= \sqrt{\left( x_{1} \right)^{2} + \left( y\left( t \right) \right)^{2}} \\ E\left( t \right)_{2} &= \sqrt{\left( x_{2} \right)^{2} + \left( y\left( t \right) \right)^{2}} \\ E\left( t \right)_{3} &= \sqrt{\left( x_{3} \right)^{2} + \left( y\left( t \right) \right)^{2}} \\ E\left( t \right)_{4} &= \sqrt{\left( x_{4} \right)^{2} + \left( y\left( t \right) \right)^{2}} \\ \mathtt{I\delta\rho}\left( t \right) \epsilon &= \left( E\left( t \right)_{1} W_{1} \psi\left( t \right)_{1,0} + E\left( t \right)_{2} W_{2} \psi\left( t \right)_{2,0} + E\left( t \right)_{3} W_{3} \psi\left( t \right)_{3,0} + E\left( t \right)_{4} W_{4} \psi\left( t \right)_{4,0} \right) \epsilon \\ \delta\left( t \right) \epsilon &= \frac{\mathtt{I\delta\rho}\left( t \right) \epsilon}{\mathtt{I\rho}\left( t \right)} \\ u\left( t \right) \epsilon &= \frac{\left( W_{1} x_{1} \psi\left( t \right)_{1,1} + W_{2} x_{2} \psi\left( t \right)_{2,1} + W_{3} x_{3} \psi\left( t \right)_{3,1} + W_{4} x_{4} \psi\left( t \right)_{4,1} \right) \epsilon}{\frac{1}{3} \mathtt{IP}\left( t \right) + \mathtt{I\rho}\left( t \right)} \\ \theta\left( t \right) \epsilon &= k u\left( t \right) \epsilon \\ \sigma\left( t \right) \epsilon &= \frac{\frac{2}{3} \left( \frac{\left( x_{4} \right)^{2} W_{4} \psi\left( t \right)_{4,2}}{E\left( t \right)_{4}} + \frac{\left( x_{3} \right)^{2} W_{3} \psi\left( t \right)_{3,2}}{E\left( t \right)_{3}} + \frac{\left( x_{2} \right)^{2} W_{2} \psi\left( t \right)_{2,2}}{E\left( t \right)_{2}} + \frac{\left( x_{1} \right)^{2} W_{1} \psi\left( t \right)_{1,2}}{E\left( t \right)_{1}} \right) \epsilon}{\frac{1}{3} \mathtt{IP}\left( t \right) + \mathtt{I\rho}\left( t \right)} \\ \mathtt{c_s^2}\left( t \right) \epsilon &= \frac{\left( \frac{\left( x_{1} \right)^{2} W_{1} \psi\left( t \right)_{1,0}}{E\left( t \right)_{1}} + \frac{\left( x_{4} \right)^{2} W_{4} \psi\left( t \right)_{4,0}}{E\left( t \right)_{4}} + \frac{\left( x_{2} \right)^{2} W_{2} \psi\left( t \right)_{2,0}}{E\left( t \right)_{2}} + \frac{\left( x_{3} \right)^{2} W_{3} \psi\left( t \right)_{3,0}}{E\left( t \right)_{3}} \right) \epsilon}{\mathtt{I\delta\rho}\left( t \right)} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{1,0} \epsilon &= \left( \frac{x_{1} \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t}}{1 + e^{ - x_{1}}} + \frac{ - k x_{1} \psi\left( t \right)_{1,1}}{E\left( t \right)_{1}} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{1,1} \epsilon &= \left( \frac{\frac{1}{3} E\left( t \right)_{1} k \Psi\left( t \right)}{1 + e^{ - x_{1}}} + \frac{\frac{1}{3} k x_{1} \left( \psi\left( t \right)_{1,0} - 2 \psi\left( t \right)_{1,2} \right)}{E\left( t \right)_{1}} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{1,2} \epsilon &= \frac{\frac{1}{5} k x_{1} \left( 2 \psi\left( t \right)_{1,1} - 3 \psi\left( t \right)_{1,3} \right) \epsilon}{E\left( t \right)_{1}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{1,3} \epsilon &= \frac{\frac{1}{7} k x_{1} \left( 3 \psi\left( t \right)_{1,2} - 4 \psi\left( t \right)_{1,4} \right) \epsilon}{E\left( t \right)_{1}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{1,4} \epsilon &= \frac{\frac{1}{9} k x_{1} \left( 4 \psi\left( t \right)_{1,3} - 5 \psi\left( t \right)_{1,5} \right) \epsilon}{E\left( t \right)_{1}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{1,5} \epsilon &= \frac{\frac{1}{11} k x_{1} \left( 5 \psi\left( t \right)_{1,4} - 6 \psi\left( t \right)_{1,6} \right) \epsilon}{E\left( t \right)_{1}} \\ \psi\left( t \right)_{1,6} \epsilon &= \left( \frac{11 E\left( t \right)_{1} \mathscr{E}\left( t \right) \psi\left( t \right)_{1,5}}{k x_{1}} - \psi\left( t \right)_{1,4} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{2,0} \epsilon &= \left( \frac{x_{2} \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t}}{1 + e^{ - x_{2}}} + \frac{ - k x_{2} \psi\left( t \right)_{2,1}}{E\left( t \right)_{2}} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{2,1} \epsilon &= \left( \frac{\frac{1}{3} E\left( t \right)_{2} k \Psi\left( t \right)}{1 + e^{ - x_{2}}} + \frac{\frac{1}{3} k x_{2} \left( \psi\left( t \right)_{2,0} - 2 \psi\left( t \right)_{2,2} \right)}{E\left( t \right)_{2}} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{2,2} \epsilon &= \frac{\frac{1}{5} k x_{2} \left( 2 \psi\left( t \right)_{2,1} - 3 \psi\left( t \right)_{2,3} \right) \epsilon}{E\left( t \right)_{2}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{2,3} \epsilon &= \frac{\frac{1}{7} k x_{2} \left( 3 \psi\left( t \right)_{2,2} - 4 \psi\left( t \right)_{2,4} \right) \epsilon}{E\left( t \right)_{2}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{2,4} \epsilon &= \frac{\frac{1}{9} k x_{2} \left( 4 \psi\left( t \right)_{2,3} - 5 \psi\left( t \right)_{2,5} \right) \epsilon}{E\left( t \right)_{2}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{2,5} \epsilon &= \frac{\frac{1}{11} k x_{2} \left( 5 \psi\left( t \right)_{2,4} - 6 \psi\left( t \right)_{2,6} \right) \epsilon}{E\left( t \right)_{2}} \\ \psi\left( t \right)_{2,6} \epsilon &= \left( \frac{11 E\left( t \right)_{2} \mathscr{E}\left( t \right) \psi\left( t \right)_{2,5}}{k x_{2}} - \psi\left( t \right)_{2,4} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{3,0} \epsilon &= \left( \frac{x_{3} \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t}}{1 + e^{ - x_{3}}} + \frac{ - k x_{3} \psi\left( t \right)_{3,1}}{E\left( t \right)_{3}} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{3,1} \epsilon &= \left( \frac{\frac{1}{3} E\left( t \right)_{3} k \Psi\left( t \right)}{1 + e^{ - x_{3}}} + \frac{\frac{1}{3} k x_{3} \left( \psi\left( t \right)_{3,0} - 2 \psi\left( t \right)_{3,2} \right)}{E\left( t \right)_{3}} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{3,2} \epsilon &= \frac{\frac{1}{5} k x_{3} \left( 2 \psi\left( t \right)_{3,1} - 3 \psi\left( t \right)_{3,3} \right) \epsilon}{E\left( t \right)_{3}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{3,3} \epsilon &= \frac{\frac{1}{7} k x_{3} \left( 3 \psi\left( t \right)_{3,2} - 4 \psi\left( t \right)_{3,4} \right) \epsilon}{E\left( t \right)_{3}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{3,4} \epsilon &= \frac{\frac{1}{9} k x_{3} \left( 4 \psi\left( t \right)_{3,3} - 5 \psi\left( t \right)_{3,5} \right) \epsilon}{E\left( t \right)_{3}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{3,5} \epsilon &= \frac{\frac{1}{11} k x_{3} \left( 5 \psi\left( t \right)_{3,4} - 6 \psi\left( t \right)_{3,6} \right) \epsilon}{E\left( t \right)_{3}} \\ \psi\left( t \right)_{3,6} \epsilon &= \left( \frac{11 E\left( t \right)_{3} \mathscr{E}\left( t \right) \psi\left( t \right)_{3,5}}{k x_{3}} - \psi\left( t \right)_{3,4} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{4,0} \epsilon &= \left( \frac{ - k x_{4} \psi\left( t \right)_{4,1}}{E\left( t \right)_{4}} + \frac{x_{4} \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t}}{1 + e^{ - x_{4}}} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{4,1} \epsilon &= \left( \frac{\frac{1}{3} k x_{4} \left( \psi\left( t \right)_{4,0} - 2 \psi\left( t \right)_{4,2} \right)}{E\left( t \right)_{4}} + \frac{\frac{1}{3} E\left( t \right)_{4} k \Psi\left( t \right)}{1 + e^{ - x_{4}}} \right) \epsilon \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{4,2} \epsilon &= \frac{\frac{1}{5} k x_{4} \left( 2 \psi\left( t \right)_{4,1} - 3 \psi\left( t \right)_{4,3} \right) \epsilon}{E\left( t \right)_{4}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{4,3} \epsilon &= \frac{\frac{1}{7} k x_{4} \left( 3 \psi\left( t \right)_{4,2} - 4 \psi\left( t \right)_{4,4} \right) \epsilon}{E\left( t \right)_{4}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{4,4} \epsilon &= \frac{\frac{1}{9} k x_{4} \left( 4 \psi\left( t \right)_{4,3} - 5 \psi\left( t \right)_{4,5} \right) \epsilon}{E\left( t \right)_{4}} \\ \frac{\mathrm{d}}{\mathrm{d}t} \psi\left( t \right)_{4,5} \epsilon &= \frac{\frac{1}{11} k x_{4} \left( 5 \psi\left( t \right)_{4,4} - 6 \psi\left( t \right)_{4,6} \right) \epsilon}{E\left( t \right)_{4}} \\ \psi\left( t \right)_{4,6} \epsilon &= \left( \frac{11 E\left( t \right)_{4} \mathscr{E}\left( t \right) \psi\left( t \right)_{4,5}}{k x_{4}} - \psi\left( t \right)_{4,4} \right) \epsilon \end{align*} \]

Initialization equations

initialization_equations(h)

\[ \begin{align*} \psi\left( t \right)_{1,0} \epsilon &= \frac{ - \frac{1}{2} x_{1} \Psi\left( t \right) \epsilon}{1 + e^{ - x_{1}}} \\ \psi\left( t \right)_{1,1} \epsilon &= \frac{0.16667 E\left( t \right)_{1} k \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right) \left( 1 + e^{ - x_{1}} \right)} \\ \psi\left( t \right)_{1,2} \epsilon &= \frac{\frac{1}{30} \left( \frac{k}{\mathscr{E}\left( t \right)} \right)^{2} x_{1} \Psi\left( t \right) \epsilon}{1 + e^{ - x_{1}}} \\ \psi\left( t \right)_{1,3} \epsilon &= 0 \\ \psi\left( t \right)_{1,4} \epsilon &= 0 \\ \psi\left( t \right)_{1,5} \epsilon &= 0 \\ \psi\left( t \right)_{2,0} \epsilon &= \frac{ - \frac{1}{2} x_{2} \Psi\left( t \right) \epsilon}{1 + e^{ - x_{2}}} \\ \psi\left( t \right)_{2,1} \epsilon &= \frac{0.16667 E\left( t \right)_{2} k \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right) \left( 1 + e^{ - x_{2}} \right)} \\ \psi\left( t \right)_{2,2} \epsilon &= \frac{\frac{1}{30} \left( \frac{k}{\mathscr{E}\left( t \right)} \right)^{2} x_{2} \Psi\left( t \right) \epsilon}{1 + e^{ - x_{2}}} \\ \psi\left( t \right)_{2,3} \epsilon &= 0 \\ \psi\left( t \right)_{2,4} \epsilon &= 0 \\ \psi\left( t \right)_{2,5} \epsilon &= 0 \\ \psi\left( t \right)_{3,0} \epsilon &= \frac{ - \frac{1}{2} x_{3} \Psi\left( t \right) \epsilon}{1 + e^{ - x_{3}}} \\ \psi\left( t \right)_{3,1} \epsilon &= \frac{0.16667 E\left( t \right)_{3} k \Psi\left( t \right) \epsilon}{\left( 1 + e^{ - x_{3}} \right) \mathscr{E}\left( t \right)} \\ \psi\left( t \right)_{3,2} \epsilon &= \frac{\frac{1}{30} \left( \frac{k}{\mathscr{E}\left( t \right)} \right)^{2} x_{3} \Psi\left( t \right) \epsilon}{1 + e^{ - x_{3}}} \\ \psi\left( t \right)_{3,3} \epsilon &= 0 \\ \psi\left( t \right)_{3,4} \epsilon &= 0 \\ \psi\left( t \right)_{3,5} \epsilon &= 0 \\ \psi\left( t \right)_{4,0} \epsilon &= \frac{ - \frac{1}{2} x_{4} \Psi\left( t \right) \epsilon}{1 + e^{ - x_{4}}} \\ \psi\left( t \right)_{4,1} \epsilon &= \frac{0.16667 E\left( t \right)_{4} k \Psi\left( t \right) \epsilon}{\left( 1 + e^{ - x_{4}} \right) \mathscr{E}\left( t \right)} \\ \psi\left( t \right)_{4,2} \epsilon &= \frac{\frac{1}{30} \left( \frac{k}{\mathscr{E}\left( t \right)} \right)^{2} x_{4} \Psi\left( t \right) \epsilon}{1 + e^{ - x_{4}}} \\ \psi\left( t \right)_{4,3} \epsilon &= 0 \\ \psi\left( t \right)_{4,4} \epsilon &= 0 \\ \psi\left( t \right)_{4,5} \epsilon &= 0 \end{align*} \]

Dark energy

SymBoltz.cosmological_constant — Function

cosmological_constant(g; name = :Λ, kwargs...)

Create a particle species for the cosmological constant (with equation of state w ~ -1) in the spacetime with metric g.

Λ = SymBoltz.cosmological_constant(g)

Equations

equations(Λ)

\[ \begin{align*} \delta\left( t \right) \epsilon &= 0 \\ \theta\left( t \right) \epsilon &= 0 \\ \sigma\left( t \right) \epsilon &= 0 \\ \mathtt{c_s^2}\left( t \right) \epsilon &= - \epsilon \\ w\left( t \right) &= -1 \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \rho\left( t \right) &= 0.11937 \left( a\left( t \right) \right)^{ - 3 \left( 1 + w\left( t \right) \right)} \mathtt{\Omega{_0}} \\ \Omega\left( t \right) &= 8.3776 \rho\left( t \right) \end{align*} \]

Initialization equations

initialization_equations(Λ)

\[ \begin{align*} \end{align*} \]

SymBoltz.w0wa — Function

w0wa(g; kwargs...)

Create a particle species for the w₀-wₐ dark energy (CPL) parametrization in the spacetime with metric g.

X = SymBoltz.w0wa(g)

Equations

equations(X)

\[ \begin{align*} w\left( t \right) &= \mathtt{w0} + \mathtt{wa} \left( 1 - a\left( t \right) \right) \\ \dot{w}\left( t \right) &= \frac{\mathrm{d} w\left( t \right)}{\mathrm{d}t} \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \frac{\mathrm{d} \rho\left( t \right)}{\mathrm{d}t} &= - 3 \mathscr{E}\left( t \right) \rho\left( t \right) \left( 1 + w\left( t \right) \right) \\ \frac{\mathrm{d} \delta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \left( \theta\left( t \right) - 3 \frac{\mathrm{d} \Phi\left( t \right)}{\mathrm{d}t} \right) \left( -1 - w\left( t \right) \right) - 3 \left( \mathtt{c_s^2} - w\left( t \right) \right) \mathscr{E}\left( t \right) \delta\left( t \right) \right) \epsilon \\ \frac{\mathrm{d} \theta\left( t \right)}{\mathrm{d}t} \epsilon &= \left( \frac{k^{2} \mathtt{c_s^2} \delta\left( t \right)}{1 + w\left( t \right)} + \frac{ - \theta\left( t \right) \frac{\mathrm{d} w\left( t \right)}{\mathrm{d}t}}{1 + w\left( t \right)} - k^{2} \sigma\left( t \right) + k^{2} \Psi\left( t \right) - \mathscr{E}\left( t \right) \theta\left( t \right) \left( 1 - 3 w\left( t \right) \right) \right) \epsilon \\ \sigma\left( t \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(X)

\[ \begin{align*} \delta\left( t \right) \epsilon &= - \frac{3}{2} \Psi\left( t \right) \left( 1 + w\left( t \right) \right) \epsilon \\ \theta\left( t \right) \epsilon &= \frac{\frac{1}{2} k^{2} \Psi\left( t \right) \epsilon}{\mathscr{E}\left( t \right)} \end{align*} \]

SymBoltz.quintessence — Function

quintessence(g, v; name = :ϕ, kwargs...)

Create a species for a quintessence scalar field with potential v in the spacetime with metric g.

@variables ϕ(t) V(ϕ) V′(ϕ) V′′(ϕ)
Q = SymBoltz.quintessence(g, V, V′, V′′)

Equations

equations(Q)

\[ \begin{align*} \frac{\mathrm{d}}{\mathrm{d}t} \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} &= - 2 \mathscr{E}\left( t \right) \frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t} - \left( a\left( t \right) \right)^{2} \mathtt{V\prime}\left( t \right) \\ V\left( t \right) &= V\left( \phi\left( t \right) \right) \\ \mathtt{V\prime}\left( t \right) &= \mathtt{V\prime}\left( \phi\left( t \right) \right) \\ \mathtt{V\prime\prime}\left( t \right) &= \mathtt{V\prime\prime}\left( \phi\left( t \right) \right) \\ K\left( t \right) &= \frac{1}{2} \left( \frac{\frac{\mathrm{d} \phi\left( t \right)}{\mathrm{d}t}}{a\left( t \right)} \right)^{2} \\ \rho\left( t \right) &= V\left( t \right) + K\left( t \right) \\ P\left( t \right) &= - V\left( t \right) + K\left( t \right) \\ w\left( t \right) &= \frac{P\left( t \right)}{\rho\left( t \right)} \\ \mathtt{m^2}\left( t \right) &= \mathtt{V\prime\prime}\left( t \right) \\ \mathtt{{\epsilon}s}\left( t \right) &= \frac{\left( \frac{\mathtt{V\prime}\left( t \right)}{V\left( t \right)} \right)^{2}}{16 \pi} \\ \mathtt{{\eta}s}\left( t \right) &= \frac{\mathtt{V\prime\prime}\left( t \right)}{8 V\left( t \right) \pi} \\ \delta\left( t \right) \epsilon &= 0 \\ \sigma\left( t \right) \epsilon &= 0 \\ \mathtt{c_s^2}\left( t \right) \epsilon &= 0 \end{align*} \]

Initialization equations

initialization_equations(Q)

\[ \begin{align*} \end{align*} \]

Curvature (effective)

SymBoltz.curvature — Function

curvature(g; name = :K, kwargs...)

Create a species that effectively accounts for curvature in the spacetime with metric g.

K = SymBoltz.curvature(g)

Equations

equations(K)

\[ \begin{align*} w\left( t \right) &= \frac{-1}{3} \\ \rho\left( t \right) &= \frac{3 \mathtt{\Omega_0}}{8 \left( a\left( t \right) \right)^{2} \pi} \\ P\left( t \right) &= \rho\left( t \right) w\left( t \right) \\ \delta\left( t \right) &= 0 \\ \theta\left( t \right) &= 0 \\ \mathtt{c_s^2}\left( t \right) &= 0 \\ \sigma\left( t \right) &= 0 \end{align*} \]

Initialization equations

initialization_equations(K)

\[ \begin{align*} \end{align*} \]