Components

Spacetimes

SymBoltz.metric — Function

metric(; name = :g, kwargs...)

Create a symbolic component for the perturbed FLRW spacetime metric in the conformal Newtonian gauge with signature

a^2 * diag(-1-2*ϵ*Ψ, +1 - 2*ϵ*Φ, +1 - 2*ϵ*Φ, +1 - 2*ϵ*Φ)

Gravitational theories

SymBoltz.general_relativity — Function

general_relativity(g; name = :G, kwargs...)

Create a symbolic component for the general relativistic (GR) theory of gravity in the spacetime with the metric g.

SymBoltz.brans_dicke — Function

brans_dicke(g; name = :G, kwargs...)

Create a symbolic component for the Brans-Dicke (BD) theory of gravity in the spacetime with the metric g.

Species

Generic

SymBoltz.species_constant_eos — Function

species_constant_eos(g, w, cₛ² = w, ẇ = 0, _σ = 0; θinteract = false, kwargs...)

Create a symbolic component for a particle species with equation of state w ~ P/ρ in the spacetime with the metric g.

Relativistic

SymBoltz.radiation — Function

radiation(g; name = :r, kwargs...)

Create a particle species for radiation (with equation of state w ~ 1/3) in the spacetime with metric g.

SymBoltz.photons — Function

photons(g; polarization = true, lmax = 6, name = :γ, kwargs...)

Create a particle species for photons in the spacetime with metric g.

Non-relativistic

SymBoltz.matter — Function

matter(g; name = :m, kwargs...)

Create a particle species for matter (with equation of state w ~ 0) in the spacetime with metric g.

SymBoltz.cold_dark_matter — Function

cold_dark_matter(g; name = :c, kwargs...)

Create a particle species for cold dark matter in the spacetime with metric g.

SymBoltz.baryons — Function

baryons(g; recombination=true, name = :b, kwargs...)

Create a particle species for baryons in the spacetime with metric g.

Neutrinos

SymBoltz.massless_neutrinos — Function

massless_neutrinos(g; lmax = 6, name = :ν, kwargs...)

Create a particle species for massless neutrinos in the spacetime with metric g.

SymBoltz.massive_neutrinos — Function

massive_neutrinos(g; nx = 5, lmax = 4, name = :h, kwargs...)

Create a particle species for massive neutrinos in the spacetime with metric g.

Dark energy

SymBoltz.cosmological_constant — Function

cosmological_constant(g; name = :Λ, kwargs...)

Create a particle species for the cosmological constant (with equation of state w ~ -1) in the spacetime with metric g.

SymBoltz.w0wa — Function

w0wa(g; kwargs...)

Create a particle species for the w₀-wₐ dark energy (CPL) parametrization in the spacetime with metric g.

SymBoltz.quintessence — Function

quintessence(g, v; name = :ϕ, kwargs...)

Create a species for a quintessence scalar field with potential v in the spacetime with metric g.