Components

SymBoltz.baryons — Method

baryons(g; recombination=true, name = :b, kwargs...)

Create a particle species for baryons in the spacetime with metric g.

SymBoltz.brans_dicke — Method

brans_dicke(g; name = :G, kwargs...)

Examples

Shoot for parameters that give E = G = 1 today:

using SymBoltz, ModelingToolkit
M = BDΛCDM()
D = Differential(M.t)

pars_fixed = [parameters_Planck18(M); M.G.ω => 100.0; D(M.G.ϕ) => 0.0] # unspecified: M.Λ.Ω0, M.G.ϕ
pars_guess = [M.G.ϕ => 0.95, M.Λ.Ω0 => 0.7] # initial guesses for shooting method
pars_shoot = shoot(M, pars_fixed, pars_guess, [M.g.ℰ ~ 1, M.G.G ~ 1]; thermo = false, backwards = false) # exact solutions
pars = [pars_fixed; pars_shoot] # merge fixed and shooting parameters

Solve background and plot scalar field and Hubble function:

using Plots
sol = solve(M, pars, thermo = false, backwards = false)
plot(sol, log10(M.g.a), [M.g.ℰ, M.G.G], ylims=(0.8, 1.2))

SymBoltz.cosmological_constant — Method

cosmological_constant(g; name = :Λ, kwargs...)

Create a particle species for the cosmological constant (with equation of state w ~ -1) in the spacetime with metric g.

SymBoltz.general_relativity — Method

general_relativity(g; name = :G, kwargs...)

Create a symbolic component for the general relativistic (GR) theory of gravity in the spacetime with the metric g.

SymBoltz.massive_neutrinos — Method

massive_neutrinos(g; nx=5, lmax=4, name = :h, kwargs...)

Create a particle species for massive neutrinos in the spacetime with metric g.

SymBoltz.massless_neutrinos — Method

massless_neutrinos(g; lmax=6, name = :ν, kwargs...)

Create a particle species for massless neutrinos in the spacetime with metric g.

SymBoltz.matter — Method

matter(g; name = :m, kwargs...)

Create a particle species for matter (with equation of state w ~ 0) in the spacetime with metric g.

SymBoltz.metric — Method

metric(; name = :g, kwargs...)

Create a symbolic component for the perturbed FLRW spacetime metric in the conformal Newtonian gauge with signature

a^2 * diag(-1-2*ϵ*Ψ, +1 - 2*ϵ*Φ, +1 - 2*ϵ*Φ, +1 - 2*ϵ*Φ)

SymBoltz.photons — Method

cosmological_constant(g; name = :Λ, kwargs...)

Create a particle species for photons in the spacetime with metric g.

SymBoltz.quintessence — Method

quintessence(g, v; name = :ϕ, kwargs...)

Create a species for a quintessence scalar field with potential v in the spacetime with metric g.

Examples

using ModelingToolkit, DifferentialEquations, Plots
@parameters V0 N
V(ϕ) = V0 * ϕ^N
M = QCDM(V)
D = Differential(M.t)
pars = [parameters_Planck18(M); M.Q.ϕ => 1; M.Q.V0 => 1e-2; M.Q.N => 2]
sol = solve(M, pars, thermo = false, solver = Tsit5(), reltol = 1e-10; guesses = [D(M.Q.ϕ) => +1.0])
plot(sol, M.Q.ϕ, M.Q.V, line_z = log10(M.g.a)) # plot V(ϕ(t))

SymBoltz.radiation — Method

radiation(g; name = :r, kwargs...)

Create a particle species for radiation (with equation of state w ~ 1/3) in the spacetime with metric g.

SymBoltz.species_constant_eos — Function

species_constant_eos(g, w, cs² = w, ẇ = 0, _σ = 0; θinteract = false, kwargs...)

Create a symbolic component for a particle species with equation of state w ~ P/ρ in the spacetime with the metric g.

SymBoltz.ΛCDM — Method

ΛCDM(;
    recombination = true,
    g = metric(),
    G = general_relativity(g),
    γ = photons(g),
    ν = massless_neutrinos(g),
    h = massive_neutrinos(g),
    c = cold_dark_matter(g; name = :c),
    b = baryons(g; recombination, name = :b),
    Λ = cosmological_constant(g),
    kwargs...
)

Create a complete ΛCDM model.