Components

SymBoltz.baryons — Method

baryons(g; recombination=true, name = :b, kwargs...)

Create a particle species for baryons in the spacetime with metric g.

SymBoltz.cosmological_constant — Method

cosmological_constant(g; name = :Λ, kwargs...)

Create a particle species for the cosmological constant (with equation of state w ~ -1) in the spacetime with metric g.

SymBoltz.general_relativity — Method

general_relativity(g; name = :G, kwargs...)

Create a symbolic component for the general relativistic (GR) theory of gravity in the spacetime with the metric g.

SymBoltz.massive_neutrinos — Method

massive_neutrinos(g; nx=5, lmax=4, name = :h, kwargs...)

Create a particle species for massive neutrinos in the spacetime with metric g.

SymBoltz.massless_neutrinos — Method

massless_neutrinos(g; lmax=6, name = :ν, kwargs...)

Create a particle species for massless neutrinos in the spacetime with metric g.

SymBoltz.matter — Method

matter(g; name = :m, kwargs...)

Create a particle species for matter (with equation of state w ~ 0) in the spacetime with metric g.

SymBoltz.metric — Method

metric(; name = :g, kwargs...)

Create a symbolic component for the perturbed FLRW spacetime metric in the conformal Newtonian gauge with signature

a^2 * diag(-1-2*ϵ*Ψ, +1 - 2*ϵ*Φ, +1 - 2*ϵ*Φ, +1 - 2*ϵ*Φ)

SymBoltz.photons — Method

cosmological_constant(g; name = :Λ, kwargs...)

Create a particle species for photons in the spacetime with metric g.

SymBoltz.quintessence — Method

quintessence(g; name = :ϕ, kwargs...)

Create a species for a quintessence scalar field in the spacetime with metric g.

Examples

M = ΛCDM()
ϕ = SymBoltz.quintessence(M.g)
M = ΛCDM(Λ = ϕ, name = :ϕCDM)
pars = [
    M.γ.T0 => 2.7
    M.c.Ω0 => 0.27
    M.b.Ω0 => 0.05
    M.ν.Neff => 3.0
    M.g.h => 0.7
    M.ϕ.M => 1e-3
    M.ϕ.α => 1.0
]
sol = solve(M, pars, reltol = 1e-8, thermo = false)

SymBoltz.radiation — Method

radiation(g; name = :r, kwargs...)

Create a particle species for radiation (with equation of state w ~ 1/3) in the spacetime with metric g.

SymBoltz.species_constant_eos — Function

species_constant_eos(g, w, cs² = w, ẇ = 0, _σ = 0; θinteract = false, kwargs...)

Create a symbolic component for a particle species with equation of state w ~ P/ρ in the spacetime with the metric g.

SymBoltz.ΛCDM — Method

ΛCDM(;
    recombination = true,
    g = metric(),
    G = general_relativity(g),
    γ = photons(g),
    ν = massless_neutrinos(g),
    h = massive_neutrinos(g),
    c = cold_dark_matter(g; name = :c),
    b = baryons(g; recombination, name = :b),
    Λ = cosmological_constant(g),
    kwargs...
)

Create a complete ΛCDM model.